Split install-stage1 test monolith by concern
Slice 2 of #297. Break tests/install-stage1.test.mjs (3,573 lines,
63 tests) into focused per-concern files to improve reviewability and
per-file test parallelism:

- install-stage1-core-test-helpers.mjs: shared preamble helpers and
  constants (spawn preload, runInstaller, fake pi/npm/tk builders,
  env scrubbing) extracted first as a foundation
- install-stage1-runtime.test.mjs (23 tests): Node minimum, private
  runtime repair, piInstalledByTlh, runtime-ownership markers,
  uninstall marker planning
- install-stage1-wrapper.test.mjs (16 tests): wrapper helper
  resolution, PATH sanitization, --pi-cmd fast path
- install-stage1-update.test.mjs (9 tests): tlh update and recovery
  update flows
- install-stage1.test.mjs (15 tests, 458 lines): remaining installer
  flags, config, and support-file tests

Pure move: test bodies, titles, and assertions are unchanged; the
combined test count stays exactly 63 with no duplicates.
