+++ tests/install-stage1.test.mjs
@@ -1,245 +1,31 @@
 import assert from "node:assert/strict";
 import { spawnSync } from "node:child_process";
-import { chmodSync, copyFileSync, existsSync, mkdirSync, mkdtempSync, readFileSync, realpathSync, rmSync, symlinkSync, writeFileSync } from "node:fs";
-import { tmpdir } from "node:os";
+import { existsSync, mkdirSync, readFileSync, rmSync } from "node:fs";
 import { delimiter, dirname, isAbsolute, join, resolve } from "node:path";
 import test from "node:test";
-import { fileURLToPath } from "node:url";
 
-import { makeTempDir, readPiLogRecords } from "./install-stage1-test-helpers.mjs";
+import { makeTempDir } from "./install-stage1-test-helpers.mjs";
+import {
+	TLH_PINNED_PI_VERSION,
+	escapeRegExp,
+	managedRtkSupportedTestPlatform,
+	pathWithoutRepoNodeModulesBin,
+	readJson,
+	readJsonLines,
+	repoRoot,
+	runHelper,
+	runStage1LocalPackageInstall,
+	scrubInstallerEnv,
+	writeFakePi,
+} from "./install-stage1-core-test-helpers.mjs";
 
 import {
-	MIN_NODE_VERSION,
-	RUNTIME_MARKER_FILENAME,
-	RUNTIME_OWNED_TOPLEVEL,
-	assertSupportedNodeRuntime,
 	buildInstallConfig,
-	nodeVersionMeetsMinimum,
 	parseArgs,
 	usage,
 } from "../scripts/tlh-install.mjs";
 import { validateInstallerTargets } from "../scripts/lib/tlh-install-paths.mjs";
 
-const repoRoot = resolve(dirname(fileURLToPath(import.meta.url)), "..");
-const repoNodeModulesBin = join(repoRoot, "node_modules", ".bin");
-const TLH_NON_PINNED_PI_VERSION = "0.80.1";
-const TLH_PINNED_PI_VERSION = "0.80.6";
-
-const TLH_PI_PACKAGE_SPEC = `@earendil-works/pi-coding-agent@${TLH_PINNED_PI_VERSION}`;
-const managedRtkSupportedTestPlatform = ["darwin", "linux"].includes(process.platform) && ["x64", "arm64"].includes(process.arch);
-
-function escapeRegExp(value) {
-	return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function pathWithoutRepoNodeModulesBin(pathValue = process.env.PATH || "") {
-	return pathValue.split(delimiter).filter((entry) => entry && resolve(entry) !== repoNodeModulesBin).join(delimiter);
-}
-
-function scrubInstallerEnv(overrides = {}, baseEnv = process.env) {
-	const env = {};
-	for (const [key, value] of Object.entries(baseEnv)) {
-		if (key === "PI_CODING_AGENT_DIR" || key.startsWith("TLH_")) continue;
-		env[key] = value;
-	}
-	return { ...env, ...overrides };
-}
-
-function runHelper(scriptRelativePath, args, { homeDir }) {
-	const scriptPath = join(repoRoot, scriptRelativePath);
-	const result = spawnSync(process.execPath, [scriptPath, ...args], {
-		cwd: repoRoot,
-		env: scrubInstallerEnv({ HOME: homeDir }),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(
-		result.status,
-		0,
-		`${scriptRelativePath} failed\nstdout:\n${result.stdout}\nstderr:\n${result.stderr}`,
-	);
-}
-
-function writeManagedRtkSpawnPreload() {
-	const dir = mkdtempSync(join(tmpdir(), "tlh-rtk-preload-"));
-	const preload = join(dir, "stub-managed-rtk-spawn.mjs");
-	writeFileSync(preload, `import { appendFileSync } from "node:fs";
-import { basename, join } from "node:path";
-import { createRequire } from "node:module";
-const require = createRequire(import.meta.url);
-const childProcess = require("node:child_process");
-const originalSpawnSync = childProcess.spawnSync;
-function optionEnv(options) {
-  return options && typeof options === "object" && options.env && typeof options.env === "object" ? options.env : {};
-}
-function targetFromArgs(args) {
-  const targetIndex = args.indexOf("--target");
-  if (targetIndex !== -1 && typeof args[targetIndex + 1] === "string" && args[targetIndex + 1]) return args[targetIndex + 1];
-  const agentDirIndex = args.indexOf("--agent-dir");
-  if (agentDirIndex !== -1 && typeof args[agentDirIndex + 1] === "string" && args[agentDirIndex + 1]) return join(args[agentDirIndex + 1], "bin", "rtk");
-  return "";
-}
-childProcess.spawnSync = function patchedSpawnSync(command, args = [], options = {}) {
-  const argv = Array.isArray(args) ? args.map((value) => String(value)) : [];
-  if (command === process.execPath && argv[0] && basename(argv[0]) === "tlh-rtk.mjs") {
-    if (process.env.TLH_TEST_RTK_LOG) {
-      appendFileSync(process.env.TLH_TEST_RTK_LOG, JSON.stringify({
-        command,
-        args: argv,
-        cwd: options.cwd || process.cwd(),
-        env: {
-          PI_CODING_AGENT_DIR: optionEnv(options).PI_CODING_AGENT_DIR,
-          PATH: optionEnv(options).PATH,
-        },
-      }) + "\\n");
-    }
-    const status = Number.parseInt(process.env.TLH_TEST_RTK_STATUS || "0", 10);
-    const stdout = process.env.TLH_TEST_RTK_STDOUT || (status === 0 ? targetFromArgs(argv) + "\\n" : "");
-    const stderr = process.env.TLH_TEST_RTK_STDERR || "";
-    return { pid: 0, output: [null, stdout, stderr], stdout, stderr, status, signal: null, error: undefined };
-  }
-  return originalSpawnSync(command, args, options);
-};
-`, "utf8");
-	return preload;
-}
-
-const managedRtkSpawnPreload = writeManagedRtkSpawnPreload();
-
-function withNodeImport(env, preloadPath) {
-	const existing = env.NODE_OPTIONS ? String(env.NODE_OPTIONS).trim() : "";
-	const importFlag = `--import=${preloadPath}`;
-	return {
-		...env,
-		NODE_OPTIONS: existing ? `${existing} ${importFlag}` : importFlag,
-	};
-}
-
-function runInstaller(args, env = scrubInstallerEnv()) {
-	return spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-install.mjs"), ...args], {
-		cwd: repoRoot,
-		env: withNodeImport(env, managedRtkSpawnPreload),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-}
-
-function writeFakeCommand(fakebin, name, body) {
-	mkdirSync(fakebin, { recursive: true });
-	const commandPath = join(fakebin, name);
-	writeFileSync(commandPath, `#!/usr/bin/env bash\nset -euo pipefail\n${body}\n`, "utf8");
-	chmodSync(commandPath, 0o755);
-}
-
-function writeFakePi(fakebin, body) {
-	writeFakeCommand(fakebin, "pi", body);
-}
-
-function readJson(path) {
-	return JSON.parse(readFileSync(path, "utf8"));
-}
-
-function safeInstallerPath(fakebin) {
-	return [fakebin, "/usr/bin", "/bin", "/usr/sbin", "/sbin"].join(delimiter);
-}
-
-function writeFakeTk(fakebin) {
-	writeFakeCommand(fakebin, "tk", "printf 'Usage: tk help\\nTicket CLI helper\\n'");
-}
-
-function writeLoggingPi(commandDir, logPath, version = TLH_PINNED_PI_VERSION) {
-	writeFakePi(commandDir, [
-		`printf '%s|%s|%s\\n' "\${PI_CODING_AGENT_DIR:-}" "$PWD" "$*" >>"${logPath}"`,
-		`if [[ "\${1:-}" == "--version" ]]; then printf '${version}\\n'; exit 0; fi`,
-		"exit 0",
-	].join("\n"));
-}
-
-function writeVersionedWrapperPi(commandDir, logPath, version = TLH_PINNED_PI_VERSION) {
-	writeFakePi(commandDir, [
-		`if [[ "\${1:-}" == "--version" ]]; then printf '${version}\\n'; exit 0; fi`,
-		`{ printf 'cmd=%s\\n' "$0"; printf 'argv=%s\\n' "$*"; printf 'agent=%s\\n' "\${PI_CODING_AGENT_DIR:-}"; printf 'path=%s\\n' "\${PATH:-}"; } >"${logPath}"`,
-		"exit 0",
-	].join("\n"));
-}
-
-function writeWrapperHelperLogger(scriptPath, logEnvVar, source) {
-	mkdirSync(dirname(scriptPath), { recursive: true });
-	writeFileSync(scriptPath, `import { writeFileSync } from "node:fs";\nwriteFileSync(process.env.${logEnvVar}, JSON.stringify({ source: ${JSON.stringify(source)}, argv: process.argv.slice(2), env: { PI_CODING_AGENT_DIR: process.env.PI_CODING_AGENT_DIR, PATH: process.env.PATH } }));\n`, "utf8");
-}
-
-function writeFakeNpmInstaller(fakebin, { npmLog, templatePiPath, installedPiPath }) {
-	writeFakeCommand(fakebin, "npm", [
-		`printf '%s\\n' "$*" >>"${npmLog}"`,
-		`mkdir -p "${dirname(installedPiPath)}"`,
-		`cp "${templatePiPath}" "${installedPiPath}"`,
-		`chmod +x "${installedPiPath}"`,
-	].join("\n"));
-}
-
-function runStage1LocalPackageInstall(t, {
-	dryRun = false,
-	noSettings = false,
-	force = false,
-	verbose = false,
-	existingSupportFiles,
-	existingLibrarianConfig,
-	envOverrides = {},
-} = {}) {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const packageDir = join(root, "package-source");
-	const piLog = join(root, "pi.log");
-	const templateDir = join(root, "pi-template");
-	const npmLog = join(root, "npm.log");
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	writeFakeTk(fakebin);
-	writeLoggingPi(fakebin, piLog);
-	// Fake npm so installPiIfNeeded never hits the network. The fake npm copies a
-	// template pi (reporting the pinned version) into the private runtime path.
-	writeLoggingPi(templateDir, piLog, TLH_PINNED_PI_VERSION);
-	writeFakeNpmInstaller(fakebin, {
-		npmLog,
-		templatePiPath: join(templateDir, "pi"),
-		installedPiPath: join(dirname(agentDir), "runtime", "bin", "pi"),
-	});
-	if (existingLibrarianConfig !== undefined) {
-		mkdirSync(join(agentDir, "extensions"), { recursive: true });
-		writeFileSync(join(agentDir, "extensions", "librarian.json"), JSON.stringify(existingLibrarianConfig, null, 2));
-	}
-	if (existingSupportFiles) {
-		for (const [relativePath, content] of Object.entries(existingSupportFiles)) {
-			const target = join(agentDir, "tlh", relativePath);
-			mkdirSync(dirname(target), { recursive: true });
-			writeFileSync(target, content);
-		}
-	}
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: `${fakebin}:${process.env.PATH || ""}`,
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-		...envOverrides,
-	});
-	const args = ["--agent-dir", agentDir, "--bin-dir", binDir, "--no-wrapper"];
-	if (dryRun) args.unshift("--dry-run");
-	if (noSettings) args.push("--no-settings");
-	if (force) args.push("--force");
-	if (verbose) args.push("--verbose");
-	const result = runInstaller(args, env);
-	return { result, homeDir, agentDir, binDir, piLog };
-}
-
-function readJsonLines(path) {
-	if (!existsSync(path)) return [];
-	return readFileSync(path, "utf8").trim().split(/\r?\n/).filter(Boolean).map((line) => JSON.parse(line));
-}
 
 test("stage-1 hides PATH-adjustment and refresh fallback detail lines unless --verbose", (t) => {
 	for (const verbose of [false, true]) {
@@ -261,336 +47,6 @@ test("stage-1 hides PATH-adjustment and refresh fallback detail lines unless --v
 	}
 });
 
-test("stage-1 enforces the TLH Node runtime minimum", () => {
-	assert.equal(MIN_NODE_VERSION, "22.19.0");
-	assert.equal(nodeVersionMeetsMinimum("22.18.9"), false);
-	assert.equal(nodeVersionMeetsMinimum("22.19.0"), true);
-	assert.equal(nodeVersionMeetsMinimum("23.0.0"), true);
-	assert.doesNotThrow(() => assertSupportedNodeRuntime("v22.19.0"));
-	assert.throws(
-		() => assertSupportedNodeRuntime("22.18.9"),
-		/Node\.js >= 22\.19\.0 is required \(found v22\.18\.9\)\. Install or upgrade Node\.js, then rerun the installer\./,
-	);
-	assert.throws(
-		() => assertSupportedNodeRuntime("not-a-version"),
-		/unable to determine Node\.js version; The Last Harness requires Node\.js >= 22\.19\.0\./,
-	);
-});
-
-test("stage-1 repairs the TLH private Pi runtime to the pinned version when it is below the pin", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const packageDir = join(root, "package-source");
-	const npmLog = join(root, "npm.log");
-	const stalePiCallLog = join(root, "stale-pi.log");
-	const repairedPiLog = join(root, "repaired-pi.log");
-	const templateDir = join(root, "pi-template");
-	const runtimeDir = join(root, "runtime");
-	const runtimeBinDir = join(runtimeDir, "bin");
-	const installedPiPath = join(runtimeBinDir, "pi");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	mkdirSync(binDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		track: "ref",
-		ref: "main",
-		packageSource: packageDir,
-		packageSourceIsDefault: false,
-		piInstalledByTlh: true,
-	}, null, 2));
-	writeFakeCommand(fakebin, "git", "exit 0");
-	writeFakeTk(fakebin);
-	// Seed a stale private runtime binary at the expected path (version 0.80.2 is below pin).
-	writeFakePi(runtimeBinDir, [
-		`printf '%s|%s|%s\\n' "\${PI_CODING_AGENT_DIR:-}" "$PWD" "$*" >>"${stalePiCallLog}"`,
-		"if [[ \"${1:-}\" == \"--version\" ]]; then printf '0.80.2\\n'; exit 0; fi",
-		"exit 0",
-	].join("\n"));
-	writeLoggingPi(templateDir, repairedPiLog, TLH_PINNED_PI_VERSION);
-	writeFakeNpmInstaller(fakebin, {
-		npmLog,
-		templatePiPath: join(templateDir, "pi"),
-		installedPiPath,
-	});
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: safeInstallerPath(fakebin),
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-	});
-	const result = runInstaller([
-		"--agent-dir", agentDir,
-		"--bin-dir", binDir,
-		"--no-settings",
-		"--no-wrapper",
-	], env);
-	const output = `${result.stdout}\n${result.stderr}`;
-
-	assert.equal(result.status, 0, output);
-	assert.match(output, new RegExp(`Repairing TLH private Pi runtime to pinned ${escapeRegExp(TLH_PINNED_PI_VERSION)}`));
-	assert.deepEqual(readFileSync(npmLog, "utf8").trim().split(/\r?\n/).filter(Boolean), [
-		`install -g --ignore-scripts --prefix ${runtimeDir} ${TLH_PI_PACKAGE_SPEC}`,
-	]);
-	// The stale pi was only probed for --version; the repaired pi is first validated
-	// for --version (post-install check) and then ran install+update.
-	assert.deepEqual(readPiLogRecords(stalePiCallLog).map((record) => record.command), ["--version"]);
-	assert.deepEqual(readPiLogRecords(repairedPiLog).map((record) => record.command), [
-		"--version",
-		`install ${packageDir}`,
-		`update ${packageDir}`,
-	]);
-	const state = readJson(join(agentDir, "tlh", "install-state.json"));
-	assert.equal(state.piInstalledByTlh, true);
-});
-
-test("stage-1 repairs the TLH private Pi runtime even when a supported Pi exists on PATH", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const supportedPiDir = join(root, "supported-pi");
-	const packageDir = join(root, "package-source");
-	const npmLog = join(root, "npm.log");
-	const pathPiLog = join(root, "path-pi.log");
-	const stalePiCallLog = join(root, "stale-pi.log");
-	const repairedPiLog = join(root, "repaired-pi.log");
-	const templateDir = join(root, "pi-template");
-	const runtimeDir = join(root, "runtime");
-	const runtimeBinDir = join(runtimeDir, "bin");
-	const installedPiPath = join(runtimeBinDir, "pi");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	mkdirSync(binDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		track: "ref",
-		ref: "main",
-		packageSource: packageDir,
-		packageSourceIsDefault: false,
-		piInstalledByTlh: true,
-	}, null, 2));
-	writeFakeCommand(fakebin, "git", "exit 0");
-	writeFakeTk(fakebin);
-	// A non-pinned PATH pi — the installer must never use it (private runtime only).
-	writeFakePi(supportedPiDir, [
-		`printf '%s|%s|%s\\n' "\${PI_CODING_AGENT_DIR:-}" "$PWD" "$*" >>"${pathPiLog}"`,
-		`if [[ "\${1:-}" == "--version" ]]; then printf '${TLH_NON_PINNED_PI_VERSION}\\n'; exit 0; fi`,
-		"exit 0",
-	].join("\n"));
-	// Stale private runtime at 0.80.7 (above pin) triggers repair.
-	writeFakePi(runtimeBinDir, [
-		`printf '%s|%s|%s\\n' "\${PI_CODING_AGENT_DIR:-}" "$PWD" "$*" >>"${stalePiCallLog}"`,
-		"if [[ \"${1:-}\" == \"--version\" ]]; then printf '0.80.7\\n'; exit 0; fi",
-		"exit 0",
-	].join("\n"));
-	writeLoggingPi(templateDir, repairedPiLog, TLH_PINNED_PI_VERSION);
-	writeFakeNpmInstaller(fakebin, {
-		npmLog,
-		templatePiPath: join(templateDir, "pi"),
-		installedPiPath,
-	});
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: [supportedPiDir, safeInstallerPath(fakebin)].join(delimiter),
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-	});
-	const result = runInstaller([
-		"--agent-dir", agentDir,
-		"--bin-dir", binDir,
-		"--no-settings",
-		"--no-wrapper",
-	], env);
-	const output = `${result.stdout}\n${result.stderr}`;
-
-	assert.equal(result.status, 0, output);
-	assert.match(output, new RegExp(`Repairing TLH private Pi runtime to pinned ${escapeRegExp(TLH_PINNED_PI_VERSION)}`));
-	assert.deepEqual(readFileSync(npmLog, "utf8").trim().split(/\r?\n/).filter(Boolean), [
-		`install -g --ignore-scripts --prefix ${runtimeDir} ${TLH_PI_PACKAGE_SPEC}`,
-	]);
-	// PATH pi must never be invoked — the installer is private-runtime-only.
-	assert.equal(existsSync(pathPiLog), false, output);
-	// Stale runtime pi was only probed for --version.
-	assert.deepEqual(readPiLogRecords(stalePiCallLog).map((record) => record.command), ["--version"]);
-	// Repaired pi is first validated for --version (post-install check) then ran install+update.
-	assert.deepEqual(readPiLogRecords(repairedPiLog).map((record) => record.command), [
-		"--version",
-		`install ${packageDir}`,
-		`update ${packageDir}`,
-	]);
-	const state = readJson(join(agentDir, "tlh", "install-state.json"));
-	assert.equal(state.piInstalledByTlh, true);
-});
-
-test("stage-1 preserves piInstalledByTlh=true when rerunning with a valid private runtime", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const packageDir = join(root, "package-source");
-	const npmLog = join(root, "npm.log");
-	const piLog = join(root, "pi.log");
-	const runtimeDir = join(root, "runtime");
-	const runtimeBinDir = join(runtimeDir, "bin");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	mkdirSync(binDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		track: "ref",
-		ref: "main",
-		packageSource: packageDir,
-		packageSourceIsDefault: false,
-		piInstalledByTlh: true,
-	}, null, 2));
-	writeFakeCommand(fakebin, "git", "exit 0");
-	// npm fails loudly if invoked — must NOT be called when runtime is already valid.
-	writeFakeCommand(fakebin, "npm", `printf '%s\\n' "$*" >>"${npmLog}"\nexit 97`);
-	writeFakeTk(fakebin);
-	// Valid private runtime at the pinned version.
-	writeLoggingPi(runtimeBinDir, piLog, TLH_PINNED_PI_VERSION);
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: safeInstallerPath(fakebin),
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-	});
-	const result = runInstaller([
-		"--agent-dir", agentDir,
-		"--bin-dir", binDir,
-		"--no-settings",
-		"--no-wrapper",
-		"--pi-installed-by-tlh", "true",
-	], env);
-	const output = `${result.stdout}\n${result.stderr}`;
-
-	assert.equal(result.status, 0, output);
-	assert.equal(existsSync(npmLog), false, output);
-	assert.deepEqual(readPiLogRecords(piLog).map((record) => record.command), [
-		"--version",
-		`install ${packageDir}`,
-		`update ${packageDir}`,
-	]);
-	const state = readJson(join(agentDir, "tlh", "install-state.json"));
-	assert.equal(state.piInstalledByTlh, true);
-});
-
-test("stage-1 records piInstalledByTlh=true when installing the private runtime", (t) => {
-	for (const scenario of [
-		{
-			name: "previous false field",
-			state: {
-				schemaVersion: 1,
-				repo: "diegopetrucci/the-last-harness",
-				track: "ref",
-				ref: "main",
-				packageSource: "git:github.com/diegopetrucci/the-last-harness@main",
-				packageSourceIsDefault: true,
-				piInstalledByTlh: false,
-			},
-			args: ["--pi-installed-by-tlh", "false"],
-		},
-		{
-			name: "missing previous field",
-			state: {
-				schemaVersion: 1,
-				repo: "diegopetrucci/the-last-harness",
-				track: "ref",
-				ref: "main",
-				packageSource: "git:github.com/diegopetrucci/the-last-harness@main",
-				packageSourceIsDefault: true,
-			},
-			args: [],
-		},
-	]) {
-		const root = makeTempDir(`tlh-install-stage1-pi-runtime-${scenario.name.replace(/\s+/g, "-")}-`);
-		const homeDir = join(root, "home");
-		const agentDir = join(root, "agent");
-		const binDir = join(root, "bin");
-		const fakebin = join(root, "fakebin");
-		const packageDir = join(root, "package-source");
-		const npmLog = join(root, "npm.log");
-		const piLog = join(root, "pi.log");
-		const templateDir = join(root, "pi-template");
-		const runtimeDir = join(root, "runtime");
-		const installedPiPath = join(runtimeDir, "bin", "pi");
-		t.after(() => rmSync(root, { recursive: true, force: true }));
-		mkdirSync(homeDir, { recursive: true });
-		mkdirSync(agentDir, { recursive: true });
-		mkdirSync(binDir, { recursive: true });
-		mkdirSync(packageDir, { recursive: true });
-		mkdirSync(join(agentDir, "tlh"), { recursive: true });
-		writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify(scenario.state, null, 2));
-		writeFakeCommand(fakebin, "git", "exit 0");
-		writeFakeTk(fakebin);
-		writeLoggingPi(templateDir, piLog, TLH_PINNED_PI_VERSION);
-		writeFakeNpmInstaller(fakebin, {
-			npmLog,
-			templatePiPath: join(templateDir, "pi"),
-			installedPiPath,
-		});
-
-		const env = scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: safeInstallerPath(fakebin),
-			TLH_PACKAGE_SOURCE: packageDir,
-			TLH_SKIP_GNOSIS_INSTALL: "1",
-		});
-		const result = runInstaller([
-			"--agent-dir", agentDir,
-			"--bin-dir", binDir,
-			"--no-settings",
-			"--no-wrapper",
-			...scenario.args,
-		], env);
-		const output = `${result.stdout}\n${result.stderr}`;
-
-		assert.equal(result.status, 0, `${scenario.name}\n${output}`);
-		assert.deepEqual(readFileSync(npmLog, "utf8").trim().split(/\r?\n/).filter(Boolean), [
-			`install -g --ignore-scripts --prefix ${runtimeDir} ${TLH_PI_PACKAGE_SPEC}`,
-		], scenario.name);
-		const state = readJson(join(agentDir, "tlh", "install-state.json"));
-		assert.equal(state.piInstalledByTlh, true, scenario.name);
-		assert.deepEqual(readPiLogRecords(piLog).map((record) => record.command), [
-			"--version",
-			`install ${packageDir}`,
-			`update ${packageDir}`,
-		], scenario.name);
-	}
-});
-
-
-
-test("declared Node minimum stays aligned across installer metadata", () => {
-	const packageJson = JSON.parse(readFileSync(join(repoRoot, "package.json"), "utf8"));
-	const installSh = readFileSync(join(repoRoot, "install.sh"), "utf8");
-	const releaseWorkflow = readFileSync(join(repoRoot, ".github/workflows/release.yml"), "utf8");
-
-	assert.equal(packageJson.engines.node, `>=${MIN_NODE_VERSION}`);
-	assert.ok(installSh.includes(`TLH_MIN_NODE_VERSION="${MIN_NODE_VERSION}"`));
-	assert.ok(releaseWorkflow.includes(`node-version: '${MIN_NODE_VERSION}'`));
-});
 
 test("stage-0 installer allowlist explicitly includes contrarian.md and web-scout.md", () => {
 	const installSh = readFileSync(join(repoRoot, "install.sh"), "utf8");
@@ -895,1090 +351,7 @@ test("stage-1 derives packageRoot from custom package source install dirs", (t) 
 	assert.equal(unsupportedConfig.packageHelperRoot, "");
 });
 
-test("wrapper skips stale fallback package helpers for unlocatable custom sources", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const updateLog = join(root, "update.json");
-	const defaultsLog = join(root, "defaults.json");
-	const ticketsLog = join(root, "tickets.json");
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		TLH_PACKAGE_SOURCE: "github:owner/repo",
-	}, {
-		...process.env,
-		PI_CODING_AGENT_DIR: join(homeDir, ".pi", "agent"),
-		TLH_AGENT_DIR: join(homeDir, ".pi", "agent"),
-		TLH_REPO: "poisoned/repo",
-		TLH_REF: "poisoned-ref",
-	});
-	const config = buildInstallConfig(parseArgs(["--agent-dir", agentDir, "--bin-dir", binDir], env), env);
-
-	writeWrapperHelperLogger(join(config.packageRoot, "scripts", "tlh-update.mjs"), "TLH_UPDATE_LOG", "stale-package");
-	writeWrapperHelperLogger(join(config.packageRoot, "scripts", "tlh-defaults.mjs"), "TLH_DEFAULTS_LOG", "stale-package");
-	writeWrapperHelperLogger(join(config.packageRoot, "scripts", "tlh-tickets.mjs"), "TLH_TICKETS_LOG", "stale-package");
-	mkdirSync(join(config.packageRoot, "config"), { recursive: true });
-	writeFileSync(join(config.packageRoot, "config", "default-extensions.json"), "[\n  \"stale-package\"\n]\n", "utf8");
-	writeWrapperHelperLogger(join(agentDir, "tlh", "recover-update.mjs"), "TLH_UPDATE_LOG", "recovery");
-	writeWrapperHelperLogger(join(agentDir, "tlh", "tlh-update.mjs"), "TLH_UPDATE_LOG", "legacy-profile");
-	writeWrapperHelperLogger(join(agentDir, "tlh", "tlh-defaults.mjs"), "TLH_DEFAULTS_LOG", "legacy-profile");
-	writeWrapperHelperLogger(join(agentDir, "tlh", "tlh-tickets.mjs"), "TLH_TICKETS_LOG", "legacy-profile");
-	writeFileSync(join(agentDir, "tlh", "default-extensions.json"), "[\n  \"legacy-profile\"\n]\n", "utf8");
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		config.agentDir,
-		"--bin-dir",
-		config.binDir,
-		"--wrapper-name",
-		config.wrapperName,
-		`--package-root=${config.packageHelperRoot}`,
-	], { homeDir });
-	const wrapper = readFileSync(config.wrapperPath, "utf8");
-	assert.ok(wrapper.split(/\r?\n/).includes("default_tlh_package_root=''"));
-
-	const wrapperEnv = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: process.env.PATH || "",
-		TLH_UPDATE_LOG: updateLog,
-		TLH_DEFAULTS_LOG: defaultsLog,
-		TLH_TICKETS_LOG: ticketsLog,
-	});
-	const wrapperPath = join(binDir, config.wrapperName);
-
-	const updateResult = spawnSync(wrapperPath, ["update", "--dry-run"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(updateResult.status, 0, updateResult.stderr);
-	const updateRecord = JSON.parse(readFileSync(updateLog, "utf8"));
-	assert.equal(updateRecord.source, "recovery");
-
-	const defaultsResult = spawnSync(wrapperPath, ["defaults", "list"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(defaultsResult.status, 1);
-	assert.match(defaultsResult.stderr, /tlh defaults package support files are missing or corrupt; run `tlh update` to recover\./);
-	assert.doesNotMatch(defaultsResult.stderr, /ERR_MODULE_NOT_FOUND/);
-	assert.equal(existsSync(defaultsLog), false);
-
-	const ticketsResult = spawnSync(wrapperPath, ["tickets", "status"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(ticketsResult.status, 1);
-	assert.match(ticketsResult.stderr, /tlh tickets package support files are missing or corrupt; run `tlh update` to recover\./);
-	assert.doesNotMatch(ticketsResult.stderr, /ERR_MODULE_NOT_FOUND/);
-	assert.equal(existsSync(ticketsLog), false);
-});
-
-test("wrapper uses original node for helpers while exposing isolated bin only for tickets and pi", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const agentBinLink = join(root, "agent-bin-link");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const fakebin = join(root, "fakebin");
-	const cwdDir = join(root, "cwd");
-	const cwdLink = join(root, "cwd-link");
-	const updateLog = join(root, "update.json");
-	const ticketsLog = join(root, "tickets.json");
-	const piLog = join(root, "pi.txt");
-	const fakeNodeLog = join(root, "fake-node.log");
-	const currentNodeLog = join(root, "current-node.log");
-	const currentPiLog = join(root, "current-pi.log");
-	const isolatedPiLog = join(root, "isolated-pi.log");
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(join(packageRoot, "scripts"), { recursive: true });
-	mkdirSync(cwdDir, { recursive: true });
-	if (process.platform !== "win32") {
-		symlinkSync(agentBin, agentBinLink, "dir");
-		symlinkSync(cwdDir, cwdLink, "dir");
-	}
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeFakeCommand(agentBin, "node", "printf 'isolated node intercepted\\n' >\"${FAKE_NODE_LOG}\"\nexit 88");
-	writeFakeCommand(agentBin, "tk", "if [[ \"${1:-}\" == \"help\" ]]; then printf 'isolated tk help\\n'; exit 0; fi\nexit 1");
-	writeFakeCommand(agentBin, "pi", "printf 'isolated pi intercepted\\n' >\"${ISOLATED_PI_LOG}\"\nexit 89");
-	writeFakeCommand(cwdDir, "node", "printf 'current-dir node intercepted\\n' >\"${CURRENT_NODE_LOG}\"\nexit 87");
-	writeFakeCommand(cwdDir, "pi", "printf 'current-dir pi intercepted\\n' >\"${CURRENT_PI_LOG}\"\nexit 86");
-	writeWrapperHelperLogger(join(agentDir, "tlh", "recover-update.mjs"), "TLH_UPDATE_LOG", "recovery");
-	writeFileSync(join(packageRoot, "scripts", "tlh-tickets.mjs"), `import { spawnSync } from "node:child_process";\nimport { writeFileSync } from "node:fs";\nconst tk = spawnSync("tk", ["help"], { encoding: "utf8" });\nwriteFileSync(process.env.TLH_TICKETS_LOG, JSON.stringify({ argv: process.argv.slice(2), env: { PI_CODING_AGENT_DIR: process.env.PI_CODING_AGENT_DIR, PATH: process.env.PATH }, tk: { status: tk.status, stdout: (tk.stdout || "").trim(), stderr: (tk.stderr || "").trim(), error: tk.error?.message } }));\nprocess.exit(tk.status ?? (tk.error ? 1 : 0));\n`, "utf8");
-	writeVersionedWrapperPi(fakebin, piLog);
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-		"--pi-cmd",
-		join(fakebin, "pi"),
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const poisonedPathEntries = ["", ".", cwdDir, agentBin];
-	if (process.platform !== "win32") poisonedPathEntries.push(cwdLink, agentBinLink);
-	poisonedPathEntries.push(fakebin, process.env.PATH || "");
-	const wrapperEnv = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: poisonedPathEntries.join(":"),
-		TLH_UPDATE_LOG: updateLog,
-		TLH_TICKETS_LOG: ticketsLog,
-		PI_WRAPPER_LOG: piLog,
-		FAKE_NODE_LOG: fakeNodeLog,
-		CURRENT_NODE_LOG: currentNodeLog,
-		CURRENT_PI_LOG: currentPiLog,
-		ISOLATED_PI_LOG: isolatedPiLog,
-	});
-
-	const updateResult = spawnSync(wrapper, ["update", "--dry-run"], {
-		cwd: cwdDir,
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(updateResult.status, 0, updateResult.stderr);
-	const updateRecord = JSON.parse(readFileSync(updateLog, "utf8"));
-	assert.deepEqual(updateRecord.argv, [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--dry-run",
-	]);
-	assert.equal(updateRecord.env.PI_CODING_AGENT_DIR, agentDir);
-	const updatePathEntries = updateRecord.env.PATH.split(":");
-	assert.equal(updatePathEntries[0], fakebin);
-	assert.equal(updatePathEntries.includes(""), false);
-	assert.equal(updatePathEntries.includes("."), false);
-	assert.equal(updatePathEntries.includes(cwdDir), false);
-	assert.equal(updatePathEntries.includes(agentBin), false);
-	if (process.platform !== "win32") {
-		assert.equal(updatePathEntries.includes(cwdLink), false);
-		assert.equal(updatePathEntries.includes(agentBinLink), false);
-	}
-	assert.equal(existsSync(currentNodeLog), false);
-	assert.equal(existsSync(fakeNodeLog), false);
-
-	const ticketsResult = spawnSync(wrapper, ["tickets", "status"], {
-		cwd: cwdDir,
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(ticketsResult.status, 0, ticketsResult.stderr);
-	const ticketsRecord = JSON.parse(readFileSync(ticketsLog, "utf8"));
-	assert.deepEqual(ticketsRecord.argv, [
-		"--settings",
-		join(agentDir, "settings.json"),
-		"--agent-dir",
-		agentDir,
-		"--wrapper-name",
-		"tlh",
-		"status",
-	]);
-	assert.equal(ticketsRecord.env.PI_CODING_AGENT_DIR, agentDir);
-	const ticketsPathEntries = ticketsRecord.env.PATH.split(":");
-	assert.deepEqual(ticketsPathEntries.slice(0, 2), [agentBin, fakebin]);
-	assert.equal(ticketsPathEntries.includes(""), false);
-	assert.equal(ticketsPathEntries.includes("."), false);
-	assert.equal(ticketsPathEntries.includes(cwdDir), false);
-	if (process.platform !== "win32") {
-		assert.equal(ticketsPathEntries.includes(cwdLink), false);
-		assert.equal(ticketsPathEntries.includes(agentBinLink), false);
-	}
-	assert.equal(ticketsRecord.tk.status, 0);
-	assert.equal(ticketsRecord.tk.stdout, "isolated tk help");
-	assert.equal(existsSync(fakeNodeLog), false);
-
-	const piResult = spawnSync(wrapper, ["chat", "--version"], {
-		cwd: cwdDir,
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(piResult.status, 0, piResult.stderr);
-	const piRecord = Object.fromEntries(readFileSync(piLog, "utf8").trim().split(/\r?\n/).map((line) => {
-		const separator = line.indexOf("=");
-		return [line.slice(0, separator), line.slice(separator + 1)];
-	}));
-	assert.equal(piRecord.argv, "chat --version");
-	assert.equal(piRecord.agent, agentDir);
-	const piPathEntries = piRecord.path.split(":");
-	assert.deepEqual(piPathEntries.slice(0, 2), [agentBin, fakebin]);
-	assert.equal(piPathEntries.includes(""), false);
-	assert.equal(piPathEntries.includes("."), false);
-	assert.equal(piPathEntries.includes(cwdDir), false);
-	if (process.platform !== "win32") {
-		assert.equal(piPathEntries.includes(cwdLink), false);
-		assert.equal(piPathEntries.includes(agentBinLink), false);
-	}
-	assert.equal(existsSync(currentPiLog), false);
-	assert.equal(existsSync(isolatedPiLog), false);
-	assert.equal(existsSync(currentNodeLog), false);
-	assert.equal(existsSync(fakeNodeLog), false);
-});
-
-test("wrapper defaults and tickets helpers do not invoke PATH pi", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const fakebin = join(root, "fakebin");
-	const defaultsLog = join(root, "defaults.json");
-	const ticketsLog = join(root, "tickets.json");
-	const piProbeLog = join(root, "pi-probe.log");
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(join(packageRoot, "scripts"), { recursive: true });
-	mkdirSync(join(packageRoot, "config"), { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeFakePi(agentBin, `printf 'managed:%s\n' "$*" >>"${piProbeLog}"
-exit 63`);
-	writeFakeCommand(agentBin, "tk", "if [[ \"${1:-}\" == \"help\" ]]; then printf 'isolated tk help\\n'; exit 0; fi\nexit 1");
-	writeFakePi(fakebin, `printf 'path:%s\n' "$*" >>"${piProbeLog}"
-exit 64`);
-	writeWrapperHelperLogger(join(packageRoot, "scripts", "tlh-defaults.mjs"), "TLH_DEFAULTS_LOG", "package");
-	writeFileSync(join(packageRoot, "config", "default-extensions.json"), "[]\n", "utf8");
-	writeFileSync(join(packageRoot, "scripts", "tlh-tickets.mjs"), `import { spawnSync } from "node:child_process";\nimport { writeFileSync } from "node:fs";\nconst tk = spawnSync("tk", ["help"], { encoding: "utf8" });\nwriteFileSync(process.env.TLH_TICKETS_LOG, JSON.stringify({ argv: process.argv.slice(2), env: { PI_CODING_AGENT_DIR: process.env.PI_CODING_AGENT_DIR, PATH: process.env.PATH }, tk: { status: tk.status, stdout: (tk.stdout || "").trim(), stderr: (tk.stderr || "").trim(), error: tk.error?.message } }));\nprocess.exit(tk.status ?? (tk.error ? 1 : 0));\n`, "utf8");
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const wrapperEnv = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: [fakebin, agentBin, process.env.PATH || ""].join(":"),
-		TLH_DEFAULTS_LOG: defaultsLog,
-		TLH_TICKETS_LOG: ticketsLog,
-	});
-
-	const defaultsResult = spawnSync(wrapper, ["defaults", "list"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(defaultsResult.status, 0, defaultsResult.stderr);
-	const defaultsRecord = JSON.parse(readFileSync(defaultsLog, "utf8"));
-	assert.deepEqual(defaultsRecord.argv, [
-		"--settings",
-		join(agentDir, "settings.json"),
-		"--defaults",
-		join(packageRoot, "config", "default-extensions.json"),
-		"list",
-	]);
-	assert.equal(defaultsRecord.env.PI_CODING_AGENT_DIR, agentDir);
-	const defaultsPathEntries = defaultsRecord.env.PATH.split(":");
-	assert.equal(defaultsPathEntries[0], fakebin);
-	assert.equal(defaultsPathEntries.includes(agentBin), false);
-	assert.equal(existsSync(piProbeLog), false);
-
-	const ticketsResult = spawnSync(wrapper, ["tickets", "status"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(ticketsResult.status, 0, ticketsResult.stderr);
-	const ticketsRecord = JSON.parse(readFileSync(ticketsLog, "utf8"));
-	assert.deepEqual(ticketsRecord.argv, [
-		"--settings",
-		join(agentDir, "settings.json"),
-		"--agent-dir",
-		agentDir,
-		"--wrapper-name",
-		"tlh",
-		"status",
-	]);
-	assert.equal(ticketsRecord.env.PI_CODING_AGENT_DIR, agentDir);
-	const ticketsPathEntries = ticketsRecord.env.PATH.split(":");
-	assert.deepEqual(ticketsPathEntries.slice(0, 2), [agentBin, fakebin]);
-	assert.equal(ticketsRecord.tk.status, 0);
-	assert.equal(ticketsRecord.tk.stdout, "isolated tk help");
-	assert.equal(existsSync(piProbeLog), false);
-});
-
-test("wrapper prefers package checkout helpers over profile copies", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const updateLog = join(root, "update.json");
-	const defaultsLog = join(root, "defaults.json");
-	const ticketsLog = join(root, "tickets.json");
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(join(packageRoot, "scripts"), { recursive: true });
-	mkdirSync(join(packageRoot, "config"), { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeWrapperHelperLogger(join(packageRoot, "scripts", "tlh-update.mjs"), "TLH_UPDATE_LOG", "package");
-	writeWrapperHelperLogger(join(packageRoot, "scripts", "tlh-defaults.mjs"), "TLH_DEFAULTS_LOG", "package");
-	writeWrapperHelperLogger(join(packageRoot, "scripts", "tlh-tickets.mjs"), "TLH_TICKETS_LOG", "package");
-	writeFileSync(join(packageRoot, "config", "default-extensions.json"), "[\n  \"package\"\n]\n", "utf8");
-	writeWrapperHelperLogger(join(agentDir, "tlh", "tlh-update.mjs"), "TLH_UPDATE_LOG", "profile");
-	writeWrapperHelperLogger(join(agentDir, "tlh", "tlh-defaults.mjs"), "TLH_DEFAULTS_LOG", "profile");
-	writeWrapperHelperLogger(join(agentDir, "tlh", "tlh-tickets.mjs"), "TLH_TICKETS_LOG", "profile");
-	writeFileSync(join(agentDir, "tlh", "default-extensions.json"), "[\n  \"profile\"\n]\n", "utf8");
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const wrapperEnv = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: process.env.PATH || "",
-		TLH_UPDATE_LOG: updateLog,
-		TLH_DEFAULTS_LOG: defaultsLog,
-		TLH_TICKETS_LOG: ticketsLog,
-	});
-
-	const updateResult = spawnSync(wrapper, ["update", "--dry-run"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(updateResult.status, 0, updateResult.stderr);
-	const updateRecord = JSON.parse(readFileSync(updateLog, "utf8"));
-	assert.equal(updateRecord.source, "package");
-	assert.deepEqual(updateRecord.argv, ["--agent-dir", agentDir, "--bin-dir", binDir, "--wrapper-name", "tlh", "--dry-run"]);
-
-	const defaultsResult = spawnSync(wrapper, ["defaults", "list"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(defaultsResult.status, 0, defaultsResult.stderr);
-	const defaultsRecord = JSON.parse(readFileSync(defaultsLog, "utf8"));
-	assert.equal(defaultsRecord.source, "package");
-	assert.deepEqual(defaultsRecord.argv, [
-		"--settings",
-		join(agentDir, "settings.json"),
-		"--defaults",
-		join(packageRoot, "config", "default-extensions.json"),
-		"list",
-	]);
-
-	const ticketsResult = spawnSync(wrapper, ["tickets", "status"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(ticketsResult.status, 0, ticketsResult.stderr);
-	const ticketsRecord = JSON.parse(readFileSync(ticketsLog, "utf8"));
-	assert.equal(ticketsRecord.source, "package");
-	assert.deepEqual(ticketsRecord.argv, [
-		"--settings",
-		join(agentDir, "settings.json"),
-		"--agent-dir",
-		agentDir,
-		"--wrapper-name",
-		"tlh",
-		"status",
-	]);
-});
-
-test("wrapper ignores stale profile defaults/tickets helpers when package helper transitive imports are corrupt", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const defaultsLog = join(root, "defaults.json");
-	const ticketsLog = join(root, "tickets.json");
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(join(packageRoot, "scripts", "lib"), { recursive: true });
-	mkdirSync(join(packageRoot, "config"), { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeWrapperHelperLogger(join(agentDir, "tlh", "tlh-defaults.mjs"), "TLH_DEFAULTS_LOG", "legacy-profile");
-	writeWrapperHelperLogger(join(agentDir, "tlh", "tlh-tickets.mjs"), "TLH_TICKETS_LOG", "legacy-profile");
-	writeFileSync(join(agentDir, "tlh", "default-extensions.json"), "[\n  \"legacy-profile\"\n]\n", "utf8");
-	writeFileSync(join(packageRoot, "scripts", "tlh-defaults.mjs"), 'import "./lib/defaults-hop.mjs";\n', "utf8");
-	writeFileSync(join(packageRoot, "scripts", "lib", "defaults-hop.mjs"), 'import "./missing-defaults-lib.mjs";\n', "utf8");
-	writeFileSync(join(packageRoot, "config", "default-extensions.json"), "[]\n", "utf8");
-	writeFileSync(join(packageRoot, "scripts", "tlh-tickets.mjs"), 'import "./lib/tickets-hop.mjs";\n', "utf8");
-	writeFileSync(join(packageRoot, "scripts", "lib", "tickets-hop.mjs"), 'import "./missing-tickets-lib.mjs";\n', "utf8");
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const wrapperEnv = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: process.env.PATH || "",
-		TLH_DEFAULTS_LOG: defaultsLog,
-		TLH_TICKETS_LOG: ticketsLog,
-	});
-
-	const defaultsResult = spawnSync(wrapper, ["defaults", "list"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(defaultsResult.status, 1);
-	assert.match(defaultsResult.stderr, /tlh defaults package support files are missing or corrupt; run `tlh update` to recover\./);
-	assert.doesNotMatch(defaultsResult.stderr, /ERR_MODULE_NOT_FOUND/);
-	assert.equal(existsSync(defaultsLog), false);
-
-	const ticketsResult = spawnSync(wrapper, ["tickets", "status"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(ticketsResult.status, 1);
-	assert.match(ticketsResult.stderr, /tlh tickets package support files are missing or corrupt; run `tlh update` to recover\./);
-	assert.doesNotMatch(ticketsResult.stderr, /ERR_MODULE_NOT_FOUND/);
-	assert.equal(existsSync(ticketsLog), false);
-});
-
-test("wrapper falls back to the profile recovery updater when package update helper transitive imports are missing", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const legacyUpdateLog = join(root, "legacy-update.log");
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(join(packageRoot, "scripts", "lib"), { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		track: "latest-release",
-		packageSource: "git:github.com/diegopetrucci/the-last-harness@main",
-		packageSourceIsDefault: true,
-	}, null, 2));
-	writeFileSync(join(packageRoot, "scripts", "tlh-update.mjs"), 'import "./lib/update-hop.mjs";\n', "utf8");
-	writeFileSync(join(packageRoot, "scripts", "lib", "update-hop.mjs"), 'import "./missing-update-hop.mjs";\n', "utf8");
-	copyFileSync(join(repoRoot, "scripts", "tlh-recover-update.mjs"), join(agentDir, "tlh", "recover-update.mjs"));
-	writeFileSync(join(agentDir, "tlh", "tlh-update.mjs"), `import { writeFileSync } from "node:fs";\nwriteFileSync(${JSON.stringify(legacyUpdateLog)}, "legacy profile update should not run\\n");\nprocess.exit(91);\n`, "utf8");
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const result = spawnSync(wrapper, ["update", "--dry-run"], {
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: process.env.PATH || "",
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-	assert.match(result.stdout, /The Last Harness update plan/);
-	assert.match(result.stdout, /releases\/latest\/download\/install\.sh/);
-	assert.equal(result.stderr, "");
-	assert.equal(existsSync(legacyUpdateLog), false);
-});
-
-test("wrapper uses the profile recovery updater before legacy profile update helpers", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const safeBin = join(root, "safe-bin");
-	const fetchPreload = join(root, "stub-recovery-fetch.mjs");
-	const bashLog = join(root, "bash.log");
-	const legacyUpdateLog = join(root, "legacy-update.log");
-	const bashPath = spawnSync("sh", ["-c", "command -v bash"], { encoding: "utf8" }).stdout.trim() || "/bin/bash";
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageRoot, { recursive: true });
-	mkdirSync(safeBin, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		track: "ref",
-		ref: "main",
-		packageSource: "git:github.com/diegopetrucci/the-last-harness@main",
-		packageSourceIsDefault: true,
-	}, null, 2));
-	copyFileSync(join(repoRoot, "scripts", "tlh-recover-update.mjs"), join(agentDir, "tlh", "recover-update.mjs"));
-	writeFileSync(join(agentDir, "tlh", "tlh-update.mjs"), `import { writeFileSync } from "node:fs";\nwriteFileSync(${JSON.stringify(legacyUpdateLog)}, "legacy profile update should not run\\n");\nprocess.exit(91);\n`, "utf8");
-	writeFileSync(join(safeBin, "bash"), [
-		"#!/bin/sh",
-		"marker=$(grep -F 'recovery stub marker' \"$1\" || true)",
-		"{ printf 'cmd=%s\\n' \"$0\"; printf 'argv=%s\\n' \"$*\"; printf 'marker=%s\\n' \"${marker}\"; printf 'repo=%s\\n' \"${TLH_REPO:-}\"; printf 'source=%s\\n' \"${TLH_PACKAGE_SOURCE:-}\"; printf 'agent=%s\\n' \"${PI_CODING_AGENT_DIR:-}\"; printf 'path=%s\\n' \"${PATH:-}\"; } >\"${BASH_LOG}\"",
-	].join("\n"), "utf8");
-	chmodSync(join(safeBin, "bash"), 0o755);
-	writeFileSync(fetchPreload, `globalThis.fetch = async () => ({\n\tok: true,\n\tstatus: 200,\n\tstatusText: "OK",\n\ttext: async () => "#!/usr/bin/env bash\\n# recovery stub marker\\nexit 0\\n",\n});\n`, "utf8");
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const result = spawnSync(bashPath, [wrapper, "update", "--quiet"], {
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: `${safeBin}:${process.env.PATH || ""}`,
-			NODE_OPTIONS: `--import=${fetchPreload}`,
-			BASH_LOG: bashLog,
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-	assert.equal(existsSync(legacyUpdateLog), false);
-	const bashRecord = Object.fromEntries(readFileSync(bashLog, "utf8").trim().split(/\r?\n/).map((line) => {
-		const separator = line.indexOf("=");
-		return [line.slice(0, separator), line.slice(separator + 1)];
-	}));
-	assert.equal(bashRecord.cmd, join(safeBin, "bash"));
-	assert.match(bashRecord.argv, new RegExp(`--agent-dir ${agentDir.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}`));
-	assert.match(bashRecord.argv, new RegExp(`--bin-dir ${binDir.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}`));
-	assert.match(bashRecord.argv, /--wrapper-name tlh/);
-	assert.match(bashRecord.argv, /--track ref/);
-	assert.match(bashRecord.argv, /--ref main/);
-	assert.match(bashRecord.argv, /--quiet/);
-	assert.equal(bashRecord.marker, "# recovery stub marker");
-	assert.equal(bashRecord.repo, "diegopetrucci/the-last-harness");
-	assert.equal(bashRecord.source, "");
-	assert.equal(bashRecord.agent, agentDir);
-});
-
-test("wrapper resolves pi to an absolute command path before exposing isolated bin", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const cwdDir = join(root, "cwd");
-	const safeBinName = "safe-bin";
-	const safeBin = join(cwdDir, safeBinName);
-	const bashEnv = join(root, "bash-env.sh");
-	const piLog = join(root, "pi.txt");
-	const isolatedPiLog = join(root, "isolated-pi.log");
-	const functionPiLog = join(root, "function-pi.log");
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageRoot, { recursive: true });
-	mkdirSync(cwdDir, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeFakePi(agentBin, "printf 'isolated pi intercepted\\n' >\"${ISOLATED_PI_LOG}\"\nexit 89");
-	writeVersionedWrapperPi(safeBin, piLog);
-	writeFileSync(bashEnv, "pi() {\n\tprintf 'shell function pi intercepted\\n' >\"${FUNCTION_PI_LOG}\"\n\treturn 79\n}\n", "utf8");
-
-	// Under the private-runtime model, --pi-cmd bakes the absolute pi path at wrapper-creation time.
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-		"--pi-cmd",
-		join(safeBin, "pi"),
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const result = spawnSync(wrapper, ["chat"], {
-		cwd: cwdDir,
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: [safeBinName, agentBin, process.env.PATH || ""].join(":"),
-			BASH_ENV: bashEnv,
-			PI_WRAPPER_LOG: piLog,
-			ISOLATED_PI_LOG: isolatedPiLog,
-			FUNCTION_PI_LOG: functionPiLog,
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-	const piRecord = Object.fromEntries(readFileSync(piLog, "utf8").trim().split(/\r?\n/).map((line) => {
-		const separator = line.indexOf("=");
-		return [line.slice(0, separator), line.slice(separator + 1)];
-	}));
-	// The absolute --pi-cmd path is exec'd directly; shell functions and PATH pi are bypassed.
-	// piRecord.cmd is $0 from the exec'd script, which is the --pi-cmd value (not necessarily realpath'd).
-	assert.ok(piRecord.cmd.endsWith("/pi"), `expected pi command to end with /pi: ${piRecord.cmd}`);
-	assert.ok(isAbsolute(piRecord.cmd), `expected absolute pi path: ${piRecord.cmd}`);
-	assert.equal(piRecord.argv, "chat");
-	assert.equal(piRecord.agent, agentDir);
-	const piPathEntries = piRecord.path.split(":");
-	// PATH for pi: managed_bin (agentBin) first, then pinned_dir (safeBin, absolute), then sanitized PATH.
-	assert.deepEqual(piPathEntries.slice(0, 2), [agentBin, safeBin]);
-	assert.equal(existsSync(isolatedPiLog), false);
-	assert.equal(existsSync(functionPiLog), false);
-});
-
-
-
-test("wrapper falls back to the sanitized PATH when HOME is unset", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const supportedPiDir = join(root, "supported-pi");
-	const piLog = join(root, "pi.txt");
-	const isolatedPiLog = join(root, "isolated-pi.log");
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageRoot, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeFakePi(agentBin, "printf 'isolated pi intercepted\\n' >\"${ISOLATED_PI_LOG}\"\nexit 89");
-	writeVersionedWrapperPi(supportedPiDir, piLog);
-
-	// In the private-runtime model, --pi-cmd is baked in at wrapper creation; HOME is not needed at runtime.
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-		"--pi-cmd",
-		join(supportedPiDir, "pi"),
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const wrapperEnv = scrubInstallerEnv({
-		PATH: [supportedPiDir, agentBin, process.env.PATH || ""].join(":"),
-		PI_WRAPPER_LOG: piLog,
-		ISOLATED_PI_LOG: isolatedPiLog,
-	});
-	delete wrapperEnv.HOME;
-	const result = spawnSync(wrapper, ["chat"], {
-		env: wrapperEnv,
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-	const piRecord = Object.fromEntries(readFileSync(piLog, "utf8").trim().split(/\r?\n/).map((line) => {
-		const separator = line.indexOf("=");
-		return [line.slice(0, separator), line.slice(separator + 1)];
-	}));
-	assert.equal(piRecord.cmd, join(supportedPiDir, "pi"));
-	assert.equal(piRecord.argv, "chat");
-	assert.equal(piRecord.agent, agentDir);
-	const piPathEntries = piRecord.path.split(":");
-	assert.deepEqual(piPathEntries.slice(0, 2), [agentBin, supportedPiDir]);
-	assert.equal(existsSync(isolatedPiLog), false);
-});
-
-test("tlh update rejects legacy ticket integration flags", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		track: "ref",
-		ref: "main",
-		packageSource: "git:github.com/diegopetrucci/the-last-harness@main",
-		packageSourceIsDefault: true,
-	}, null, 2));
-
-	const runUpdate = (...extraArgs) => spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-update.mjs"), "--dry-run", "--agent-dir", agentDir, "--bin-dir", binDir, ...extraArgs], {
-		cwd: repoRoot,
-		env: scrubInstallerEnv({ HOME: homeDir }),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	const defaultResult = runUpdate();
-	assert.equal(defaultResult.status, 0, defaultResult.stderr);
-	assert.doesNotMatch(defaultResult.stdout, /--with-tickets|--without-tickets|--no-tickets/);
-
-	for (const flag of ["--with-tickets", "--without-tickets", "--no-tickets"]) {
-		const result = runUpdate(flag);
-		assert.notEqual(result.status, 0, `expected ${flag} to be rejected`);
-		assert.match(result.stderr, new RegExp(`Unknown option for tlh update: ${flag}`));
-		assert.equal(result.stdout, "");
-	}
-});
-
-test("tlh update --extensions dry-run prints the isolated package update plan and rejects installer-only flags", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const fakebin = join(root, "fakebin");
-	const dryRunPiLog = join(root, "dry-run-pi.log");
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	writeFakePi(fakebin, "printf 'pi should not run during dry-run\\n' >\"${DRY_RUN_PI_LOG}\"\nexit 91");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	const runUpdate = (...extraArgs) => spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-update.mjs"), "--extensions", "--dry-run", "--agent-dir", agentDir, ...extraArgs], {
-		cwd: repoRoot,
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: `${fakebin}:${process.env.PATH || ""}`,
-			DRY_RUN_PI_LOG: dryRunPiLog,
-			PI_CODING_AGENT_DIR: join(homeDir, ".pi", "agent"),
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	const defaultResult = runUpdate();
-	assert.equal(defaultResult.status, 0, defaultResult.stderr);
-	assert.match(defaultResult.stdout, /The Last Harness extension update plan/);
-	assert.ok(defaultResult.stdout.includes(`Agent dir: ${agentDir}`));
-	assert.match(defaultResult.stdout, /Would run: PI_CODING_AGENT_DIR='/);
-	assert.match(defaultResult.stdout, /'update' '--extensions'/);
-	assert.equal(defaultResult.stdout.includes(join(homeDir, ".pi", "agent")), false);
-	if (managedRtkSupportedTestPlatform) {
-		assert.match(defaultResult.stderr, /Would install RTK/);
-		assert.match(defaultResult.stderr, new RegExp(escapeRegExp(join(agentDir, "bin", "rtk"))));
-	} else {
-		assert.equal(defaultResult.stderr, "");
-	}
-	assert.equal(existsSync(dryRunPiLog), false);
-
-	const unsupportedFlags = [
-		["--track", "ref"],
-		["--ref", "main"],
-		["--repo", "owner/repo"],
-		["--package-source", "git:github.com/owner/repo@main"],
-		["--force"],
-		["--no-settings"],
-		["--no-wrapper"],
-	];
-	for (const [flag, value] of unsupportedFlags) {
-		const result = value ? runUpdate(flag, value) : runUpdate(flag);
-		assert.notEqual(result.status, 0, `expected ${flag} to be rejected`);
-		assert.match(result.stderr, /--extensions does not support /);
-		assert.match(result.stderr, new RegExp(flag.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")));
-		assert.equal(result.stdout, "");
-	}
-});
-
-test("tlh update --extensions refuses to target normal Pi config via explicit or inherited agent dir selection", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const protectedAgentDir = join(homeDir, ".pi", "agent");
-	mkdirSync(homeDir, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	const scenarios = [
-		{
-			name: "explicit --agent-dir",
-			args: ["--agent-dir", protectedAgentDir],
-			env: {},
-		},
-		{
-			name: "PI_CODING_AGENT_DIR fallback",
-			args: [],
-			env: { PI_CODING_AGENT_DIR: protectedAgentDir },
-		},
-		{
-			name: "TLH_AGENT_DIR override",
-			args: [],
-			env: { TLH_AGENT_DIR: protectedAgentDir },
-		},
-	];
-
-	for (const scenario of scenarios) {
-		const result = spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-update.mjs"), "--extensions", "--dry-run", ...scenario.args], {
-			cwd: repoRoot,
-			env: scrubInstallerEnv({
-				HOME: homeDir,
-				PATH: "",
-				...scenario.env,
-			}),
-			encoding: "utf8",
-			stdio: ["ignore", "pipe", "pipe"],
-		});
-
-		assert.notEqual(result.status, 0, `expected ${scenario.name} to be rejected`);
-		assert.equal(result.stdout, "");
-		assert.match(result.stderr, /refusing to run The Last Harness extension update against normal Pi config root/);
-		assert.ok(result.stderr.includes(protectedAgentDir), `${scenario.name} stderr should mention the protected agent dir`);
-		assert.doesNotMatch(result.stderr, /required command not found on sanitized PATH: pi/);
-	}
-	assert.equal(existsSync(join(homeDir, ".pi")), false);
-});
-
-test("tlh recovery update refuses to target normal Pi config before reading install-state", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const protectedAgentDir = join(homeDir, ".pi", "agent");
-	mkdirSync(join(protectedAgentDir, "tlh"), { recursive: true });
-	writeFileSync(join(protectedAgentDir, "tlh", "install-state.json"), "{ not valid json\n", "utf8");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	const scenarios = [
-		{
-			name: "explicit --agent-dir",
-			args: ["--agent-dir", protectedAgentDir],
-			env: {},
-		},
-		{
-			name: "PI_CODING_AGENT_DIR fallback",
-			args: [],
-			env: { PI_CODING_AGENT_DIR: protectedAgentDir },
-		},
-		{
-			name: "TLH_AGENT_DIR override",
-			args: [],
-			env: { TLH_AGENT_DIR: protectedAgentDir },
-		},
-	];
-
-	for (const scenario of scenarios) {
-		const result = spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-recover-update.mjs"), "--dry-run", ...scenario.args], {
-			cwd: repoRoot,
-			env: scrubInstallerEnv({
-				HOME: homeDir,
-				PATH: "",
-				...scenario.env,
-			}),
-			encoding: "utf8",
-			stdio: ["ignore", "pipe", "pipe"],
-		});
-
-		assert.notEqual(result.status, 0, `expected ${scenario.name} to be rejected`);
-		assert.equal(result.stdout, "");
-		assert.match(result.stderr, /refusing to recover The Last Harness update against normal Pi config root/);
-		assert.ok(result.stderr.includes(protectedAgentDir), `${scenario.name} stderr should mention the protected agent dir`);
-		assert.doesNotMatch(result.stderr, /could not read .*install-state\.json/);
-		assert.doesNotMatch(result.stderr, /does not contain usable The Last Harness update metadata/);
-		assert.doesNotMatch(result.stderr, /Could not determine update track/);
-	}
-});
-
-test("tlh update --extensions uses the absolute private runtime pi binary and does not fall back to PATH", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const agentBinLink = join(root, "agent-bin-link");
-	const cwdDir = join(root, "cwd");
-	const cwdLink = join(root, "cwd-link");
-	const runtimeBinDir = join(root, "runtime", "bin"); // dirname(agentDir)/runtime/bin
-	const pathPiDir = join(root, "path-pi"); // PATH-based pi that must NOT be called
-	const piLog = join(root, "pi.txt");
-	const pathPiLog = join(root, "path-pi.log");
-	const isolatedPiLog = join(root, "isolated-pi.log");
-	const rtkLog = join(root, "rtk.jsonl");
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(cwdDir, { recursive: true });
-	if (process.platform !== "win32") {
-		symlinkSync(agentBin, agentBinLink, "dir");
-		symlinkSync(cwdDir, cwdLink, "dir");
-	}
-	// Private runtime pi at the absolute path tlh-update.mjs now uses directly.
-	writeVersionedWrapperPi(runtimeBinDir, piLog);
-	// Poisoned pi entries that must not be invoked.
-	writeFakePi(agentBin, `printf 'isolated pi intercepted\\n' >"\${ISOLATED_PI_LOG}"\nexit 89`);
-	writeFakePi(pathPiDir, `printf 'PATH pi intercepted\\n' >"\${PATH_PI_LOG}"\nexit 97`);
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	const poisonedPathEntries = ["", ".", cwdDir, agentBin];
-	if (process.platform !== "win32") poisonedPathEntries.push(cwdLink, agentBinLink);
-	poisonedPathEntries.push(pathPiDir, process.env.PATH || "");
-	const result = spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-update.mjs"), "--extensions", "--agent-dir", agentDir, "--quiet"], {
-		cwd: cwdDir,
-		env: withNodeImport(scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: poisonedPathEntries.join(":"),
-			PI_WRAPPER_LOG: piLog,
-			PATH_PI_LOG: pathPiLog,
-			ISOLATED_PI_LOG: isolatedPiLog,
-			PI_CODING_AGENT_DIR: join(homeDir, ".pi", "agent"),
-			TLH_TEST_RTK_LOG: rtkLog,
-		}), managedRtkSpawnPreload),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, result.stderr);
-	assert.equal(result.stdout, "");
-	const piRecord = Object.fromEntries(readFileSync(piLog, "utf8").trim().split(/\r?\n/).map((line) => {
-		const separator = line.indexOf("=");
-		return [line.slice(0, separator), line.slice(separator + 1)];
-	}));
-	// Must use the absolute private runtime pi — not anything resolved from PATH.
-	assert.equal(piRecord.cmd, join(runtimeBinDir, "pi"));
-	assert.equal(piRecord.argv, "update --extensions");
-	assert.equal(piRecord.agent, agentDir);
-	assert.notEqual(piRecord.agent, join(homeDir, ".pi", "agent"));
-	// PATH-based and isolated pi must not have been called.
-	assert.equal(existsSync(pathPiLog), false);
-	assert.equal(existsSync(isolatedPiLog), false);
-	if (managedRtkSupportedTestPlatform) {
-		const rtkCalls = readJsonLines(rtkLog);
-		assert.equal(rtkCalls.length, 1);
-		assert.deepEqual(rtkCalls[0]?.args, [join(repoRoot, "scripts", "tlh-rtk.mjs"), "--agent-dir", agentDir, "--target", join(agentDir, "bin", "rtk"), "--quiet", "install-managed"]);
-		assert.equal(rtkCalls[0]?.env?.PI_CODING_AGENT_DIR, agentDir);
-	} else {
-		assert.equal(existsSync(rtkLog), false);
-	}
-});
-
-test("tlh update removes isolated bin and skips non-file bash candidates before running bash", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const agentBinLink = join(root, "agent-bin-link");
-	const binDir = join(root, "bin");
-	const poisonedBin = join(root, "poisoned-bin");
-	const safeBin = join(root, "safe-bin");
-	const cwdDir = join(root, "cwd");
-	const cwdLink = join(root, "cwd-link");
-	const bashLog = join(root, "bash.txt");
-	const currentBashLog = join(root, "current-bash.log");
-	const interceptedBashLog = join(root, "intercepted-bash.log");
-	const fetchPreload = join(root, "stub-update-fetch.mjs");
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(poisonedBin, { recursive: true });
-	mkdirSync(safeBin, { recursive: true });
-	mkdirSync(cwdDir, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	if (process.platform !== "win32") {
-		symlinkSync(agentBin, agentBinLink, "dir");
-		symlinkSync(cwdDir, cwdLink, "dir");
-	}
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		track: "ref",
-		ref: "main",
-		packageSource: "git:github.com/diegopetrucci/the-last-harness@main",
-		packageSourceIsDefault: true,
-	}, null, 2));
-	writeFileSync(join(agentBin, "bash"), "#!/bin/sh\nprintf 'isolated bash intercepted\\n' >\"${INTERCEPTED_BASH_LOG}\"\nexit 88\n", "utf8");
-	chmodSync(join(agentBin, "bash"), 0o755);
-	writeFileSync(join(cwdDir, "bash"), "#!/bin/sh\nprintf 'current-dir bash intercepted\\n' >\"${CURRENT_BASH_LOG}\"\nexit 87\n", "utf8");
-	chmodSync(join(cwdDir, "bash"), 0o755);
-	mkdirSync(join(poisonedBin, "bash"), { recursive: true });
-	chmodSync(join(poisonedBin, "bash"), 0o755);
-	writeFileSync(join(safeBin, "bash"), "#!/bin/sh\n{ printf 'cmd=%s\\n' \"$0\"; printf 'argv=%s\\n' \"$*\"; printf 'path=%s\\n' \"${PATH:-}\"; } >\"${BASH_LOG}\"\n", "utf8");
-	chmodSync(join(safeBin, "bash"), 0o755);
-	writeFileSync(fetchPreload, `globalThis.fetch = async () => ({\n\tok: true,\n\tstatus: 200,\n\tstatusText: "OK",\n\ttext: async () => "#!/usr/bin/env bash\\nexit 0\\n",\n});\n`, "utf8");
-
-	const poisonedPathEntries = ["", ".", cwdDir, agentBin];
-	if (process.platform !== "win32") poisonedPathEntries.push(cwdLink, agentBinLink);
-	poisonedPathEntries.push(poisonedBin, safeBin, process.env.PATH || "");
-	const result = spawnSync(process.execPath, ["--import", fetchPreload, join(repoRoot, "scripts/tlh-update.mjs"), "--agent-dir", agentDir, "--bin-dir", binDir, "--quiet"], {
-		cwd: cwdDir,
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: poisonedPathEntries.join(":"),
-			BASH_LOG: bashLog,
-			CURRENT_BASH_LOG: currentBashLog,
-			INTERCEPTED_BASH_LOG: interceptedBashLog,
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, result.stderr);
-	assert.equal(existsSync(currentBashLog), false);
-	assert.equal(existsSync(interceptedBashLog), false);
-	const bashRecord = Object.fromEntries(readFileSync(bashLog, "utf8").trim().split(/\r?\n/).map((line) => {
-		const separator = line.indexOf("=");
-		return [line.slice(0, separator), line.slice(separator + 1)];
-	}));
-	assert.equal(bashRecord.cmd, join(safeBin, "bash"));
-	assert.match(bashRecord.argv, /--agent-dir/);
-	assert.match(bashRecord.argv, /--bin-dir/);
-	const bashPathEntries = bashRecord.path.split(":");
-	assert.equal(bashPathEntries[0], poisonedBin);
-	assert.equal(bashPathEntries[1], safeBin);
-	assert.equal(bashPathEntries.includes(""), false);
-	assert.equal(bashPathEntries.includes("."), false);
-	assert.equal(bashPathEntries.includes(cwdDir), false);
-	assert.equal(bashPathEntries.includes(agentBin), false);
-	if (process.platform !== "win32") {
-		assert.equal(bashPathEntries.includes(cwdLink), false);
-		assert.equal(bashPathEntries.includes(agentBinLink), false);
-	}
-});
-
+// (tlh update / tlh recovery update tests moved to install-stage1-update.test.mjs)
 
 test("stage-1 canonicalizes relative target dirs before deriving wrapper and state paths", (t) => {
 	const root = makeTempDir();
@@ -2082,1492 +455,4 @@ test("stage-1 canonicalizes relative target dirs before deriving wrapper and sta
 	assert.notEqual(state.agentDir, parsed.agentDirInput);
 	assert.notEqual(state.agentDir, normalPiAgentIfLeftRelative);
 	assert.equal(existsSync(join(homeDir, ".pi")), false);
-});
-
-test("wrapper --pi-cmd validates the pinned binary before fast-path exec", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const pinnedPiDir = join(root, "pinned-pi");
-	const piCallLog = join(root, "pi-calls.log");
-	const piMainLog = join(root, "pi-main.txt");
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageRoot, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeFakePi(pinnedPiDir, [
-		`printf '%s\\n' "$*" >>"${piCallLog}"`,
-		`{ printf 'cmd=%s\\n' "$0"; printf 'argv=%s\\n' "$*"; printf 'agent=%s\\n' "\${PI_CODING_AGENT_DIR:-}"; printf 'path=%s\\n' "\${PATH:-}"; } >"${piMainLog}"`,
-		"exit 0",
-	].join("\n"));
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-		"--pi-cmd",
-		join(pinnedPiDir, "pi"),
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const result = spawnSync(wrapper, ["chat"], {
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: [agentBin, process.env.PATH || ""].join(":"),
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-
-	const mainRecord = Object.fromEntries(readFileSync(piMainLog, "utf8").trim().split(/\r?\n/).map((line) => {
-		const separator = line.indexOf("=");
-		return [line.slice(0, separator), line.slice(separator + 1)];
-	}));
-	assert.equal(mainRecord.argv, "chat");
-	assert.equal(mainRecord.agent, agentDir);
-
-	const allCalls = readFileSync(piCallLog, "utf8").trim().split(/\r?\n/).filter(Boolean);
-	assert.deepEqual(allCalls, ["chat"]);
-});
-test("wrapper --pi-cmd hard-fails when the pinned path is non-executable", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const piLog = join(root, "pi.txt");
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageRoot, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	const missingPiCmd = join(root, "nonexistent", "pi");
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-		"--pi-cmd",
-		missingPiCmd,
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const result = spawnSync(wrapper, ["chat"], {
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: [agentBin, process.env.PATH || ""].join(":"),
-			PI_WRAPPER_LOG: piLog,
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	// Under the private-runtime model, a missing pinned binary is a hard error.
-	assert.equal(result.status, 1, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-	assert.match(result.stderr, /private pi runtime not found at/);
-	assert.equal(existsSync(piLog), false);
-});
-test("wrapper --pi-cmd fast path exports PATH as managed_bin:pinned_dir:sanitized_path", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const pinnedPiDir = join(root, "pinned-pi");
-	const stalePiDir = join(root, "stale-pi");
-	const piLog = join(root, "pi.txt");
-	const stalePiLog = join(root, "stale-pi.log");
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageRoot, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	// Pinned pi logs PATH so we can verify ordering.
-	writeFakePi(pinnedPiDir, [
-		`{ printf 'cmd=%s\\n' "$0"; printf 'argv=%s\\n' "$*"; printf 'agent=%s\\n' "\${PI_CODING_AGENT_DIR:-}"; printf 'path=%s\\n' "\${PATH:-}"; } >"${piLog}"`,
-		"exit 0",
-	].join("\n"));
-
-	// Stale pi should not be called and should not appear before pinned_dir.
-	writeFakePi(stalePiDir, [
-		`printf 'stale pi intercepted\\n' >"${stalePiLog}"`,
-		"exit 85",
-	].join("\n"));
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-		"--pi-cmd",
-		join(pinnedPiDir, "pi"),
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const result = spawnSync(wrapper, ["chat"], {
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: [stalePiDir, agentBin, process.env.PATH || ""].join(":"),
-			STALE_PI_LOG: stalePiLog,
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-
-	const piRecord = Object.fromEntries(readFileSync(piLog, "utf8").trim().split(/\r?\n/).map((line) => {
-		const separator = line.indexOf("=");
-		return [line.slice(0, separator), line.slice(separator + 1)];
-	}));
-	assert.equal(piRecord.argv, "chat");
-	assert.equal(piRecord.agent, agentDir);
-
-	// PATH must be: managed_bin : pinned_dir : sanitized_path (stale entries come after).
-	const piPathEntries = piRecord.path.split(":");
-	assert.equal(piPathEntries[0], agentBin, `expected managed_bin first; got ${piPathEntries.join(":")}`);
-	assert.equal(piPathEntries[1], pinnedPiDir, `expected pinned_dir second; got ${piPathEntries.join(":")}`);
-	const stalePiIndex = piPathEntries.indexOf(stalePiDir);
-	assert.ok(stalePiIndex === -1 || stalePiIndex > 1, `stale entry must not appear before pinned_dir; PATH: ${piPathEntries.join(":")}`);
-
-	// Verify stale pi was not invoked.
-	assert.equal(existsSync(stalePiLog), false);
-});
-
-test("wrapper update --extensions helper prepends executable --pi-cmd directory before the sanitized PATH", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const pinnedPiDir = join(root, "pinned-pi");
-	const cwdDir = join(root, "cwd");
-	const updateLog = join(root, "update.json");
-	const nodeDir = dirname(process.execPath);
-	const bashPath = spawnSync("sh", ["-c", "command -v bash"], { encoding: "utf8" }).stdout.trim() || "/bin/bash";
-	const bashDir = dirname(bashPath);
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageRoot, { recursive: true });
-	mkdirSync(cwdDir, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(pinnedPiDir, { recursive: true });
-	writeFileSync(join(pinnedPiDir, "pi"), `#!/bin/sh\nif [ "\${1:-}" = "--version" ]; then printf '${TLH_NON_PINNED_PI_VERSION}\\n'; exit 0; fi\nprintf 'unexpected args: %s\\n' "$*" >&2\nexit 1\n`, "utf8");
-	chmodSync(join(pinnedPiDir, "pi"), 0o755);
-	writeFileSync(join(agentDir, "tlh", "recover-update.mjs"), `import { spawnSync } from "node:child_process";\nimport { writeFileSync } from "node:fs";\nconst pi = spawnSync("pi", ["--version"], { encoding: "utf8" });\nwriteFileSync(process.env.TLH_UPDATE_LOG, JSON.stringify({ argv: process.argv.slice(2), env: { PI_CODING_AGENT_DIR: process.env.PI_CODING_AGENT_DIR, PATH: process.env.PATH }, pi: { status: pi.status, stdout: pi.stdout, stderr: pi.stderr, error: pi.error?.message } }));\nprocess.exit(pi.status ?? (pi.error ? 1 : 0));\n`, "utf8");
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-		"--pi-cmd",
-		join(pinnedPiDir, "pi"),
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const result = spawnSync(wrapper, ["update", "--extensions", "--dry-run"], {
-		cwd: cwdDir,
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: ["", ".", cwdDir, agentBin, nodeDir, bashDir].join(delimiter),
-			TLH_UPDATE_LOG: updateLog,
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-
-	const updateRecord = JSON.parse(readFileSync(updateLog, "utf8"));
-	assert.deepEqual(updateRecord.argv, [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--extensions",
-		"--dry-run",
-	]);
-	assert.equal(updateRecord.env.PI_CODING_AGENT_DIR, agentDir);
-	assert.equal(updateRecord.pi.status, 0, JSON.stringify(updateRecord));
-	assert.match(updateRecord.pi.stdout, new RegExp(escapeRegExp(TLH_NON_PINNED_PI_VERSION)));
-
-	const updatePathEntries = updateRecord.env.PATH.split(delimiter);
-	assert.equal(updatePathEntries[0], pinnedPiDir, `expected pinned_dir first; got ${updatePathEntries.join(delimiter)}`);
-	assert.equal(updatePathEntries[1], nodeDir, `expected sanitized PATH to follow pinned_dir; got ${updatePathEntries.join(delimiter)}`);
-	assert.equal(updatePathEntries.includes(""), false);
-	assert.equal(updatePathEntries.includes("."), false);
-	assert.equal(updatePathEntries.includes(cwdDir), false);
-	assert.equal(updatePathEntries.includes(agentBin), false);
-});
-
-test("wrapper update --extensions helper prepends the pinned private runtime dir to PATH", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const pinnedPiDir = join(root, "pinned-pi");
-	const cwdDir = join(root, "cwd");
-	const updateLog = join(root, "update.json");
-	const pinnedPiCallLog = join(root, "pinned-pi-calls.log");
-	const nodeDir = dirname(process.execPath);
-	const bashPath = spawnSync("sh", ["-c", "command -v bash"], { encoding: "utf8" }).stdout.trim() || "/bin/bash";
-	const bashDir = dirname(bashPath);
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageRoot, { recursive: true });
-	mkdirSync(cwdDir, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	// Pinned pi at 0.80.6 — still prepended for --extensions.
-	writeFakePi(pinnedPiDir, [
-		`printf '%s\\n' "$*" >>"${pinnedPiCallLog}"`,
-		`if [[ "\${1:-}" == "--version" ]]; then printf '0.80.6\\n'; exit 0; fi`,
-		"exit 85",
-	].join("\n"));
-	writeFileSync(join(agentDir, "tlh", "recover-update.mjs"), `import { spawnSync } from "node:child_process";\nimport { writeFileSync } from "node:fs";\nconst pi = spawnSync("pi", ["--version"], { encoding: "utf8" });\nwriteFileSync(process.env.TLH_UPDATE_LOG, JSON.stringify({ argv: process.argv.slice(2), env: { PI_CODING_AGENT_DIR: process.env.PI_CODING_AGENT_DIR, PATH: process.env.PATH }, pi: { status: pi.status, stdout: pi.stdout, stderr: pi.stderr, error: pi.error?.message } }));\nprocess.exit(pi.status ?? (pi.error ? 1 : 0));\n`, "utf8");
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-		"--pi-cmd",
-		join(pinnedPiDir, "pi"),
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const result = spawnSync(wrapper, ["update", "--extensions", "--dry-run"], {
-		cwd: cwdDir,
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: ["", ".", cwdDir, agentBin, nodeDir, bashDir].join(delimiter),
-			TLH_UPDATE_LOG: updateLog,
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	// For --extensions, the wrapper prepends the private runtime dir regardless of version.
-	assert.equal(result.status, 0, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-	const updateRecord = JSON.parse(readFileSync(updateLog, "utf8"));
-	const updatePathEntries = updateRecord.env.PATH.split(delimiter);
-	assert.equal(updatePathEntries[0], pinnedPiDir, `expected pinned pi dir first for --extensions; got ${updatePathEntries.join(delimiter)}`);
-	assert.equal(updatePathEntries.includes(""), false);
-	assert.equal(updatePathEntries.includes("."), false);
-	assert.equal(updatePathEntries.includes(cwdDir), false);
-	assert.equal(updatePathEntries.includes(agentBin), false);
-	// The update script invokes pi --version and finds the pinned runtime.
-	assert.match(updateRecord.pi.stdout, new RegExp(escapeRegExp(TLH_PINNED_PI_VERSION)));
-});
-
-test("tlh update --extensions fails when managed RTK installation fails after package update", { skip: !managedRtkSupportedTestPlatform }, (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const runtimeBinDir = join(root, "runtime", "bin");
-	const piLog = join(root, "pi.txt");
-	const rtkLog = join(root, "rtk.jsonl");
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeVersionedWrapperPi(runtimeBinDir, piLog);
-	const result = spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-update.mjs"), "--extensions", "--agent-dir", agentDir, "--quiet"], {
-		cwd: repoRoot,
-		env: withNodeImport(scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: process.env.PATH || "",
-			PI_WRAPPER_LOG: piLog,
-			TLH_TEST_RTK_LOG: rtkLog,
-			TLH_TEST_RTK_STATUS: "23",
-			TLH_TEST_RTK_STDERR: "tlh-rtk: installed RTK binary did not validate\n",
-		}), managedRtkSpawnPreload),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.notEqual(result.status, 0);
-	assert.match(result.stderr, /installed RTK binary did not validate/);
-	const rtkCalls = readJsonLines(rtkLog);
-	assert.equal(rtkCalls.length, 1);
-	assert.deepEqual(rtkCalls[0]?.args, [join(repoRoot, "scripts", "tlh-rtk.mjs"), "--agent-dir", agentDir, "--target", join(agentDir, "bin", "rtk"), "--quiet", "install-managed"]);
-});
-
-test("tlh recovery update --extensions runs managed RTK installation after the package update", { skip: !managedRtkSupportedTestPlatform }, (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const runtimeBinDir = join(root, "runtime", "bin");
-	const piLog = join(root, "pi.txt");
-	const rtkLog = join(root, "rtk.jsonl");
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	writeVersionedWrapperPi(runtimeBinDir, piLog);
-	const result = spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-recover-update.mjs"), "--extensions", "--agent-dir", agentDir, "--quiet"], {
-		cwd: repoRoot,
-		env: withNodeImport(scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: process.env.PATH || "",
-			PI_WRAPPER_LOG: piLog,
-			TLH_TEST_RTK_LOG: rtkLog,
-		}), managedRtkSpawnPreload),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, result.stderr);
-	const rtkCalls = readJsonLines(rtkLog);
-	assert.equal(rtkCalls.length, 1);
-	assert.deepEqual(rtkCalls[0]?.args, [join(repoRoot, "scripts", "tlh-rtk.mjs"), "--agent-dir", agentDir, "--target", join(agentDir, "bin", "rtk"), "--quiet", "install-managed"]);
-	assert.equal(rtkCalls[0]?.env?.PI_CODING_AGENT_DIR, agentDir);
-});
-
-test("tlh update --extensions uses absolute private runtime pi and hard-fails when missing", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const runtimeBinDir = join(root, "runtime", "bin"); // dirname(agentDir)/runtime/bin
-	const pathPiDir = join(root, "path-pi"); // PATH-based pi that must NOT be used
-	const piLog = join(root, "pi.txt");
-	const pathPiLog = join(root, "path-pi.log");
-	const rtkLog = join(root, "rtk.jsonl");
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	// Case 1: private runtime pi exists — must be used, not the PATH pi.
-	writeVersionedWrapperPi(runtimeBinDir, piLog);
-	writeFakePi(pathPiDir, `printf 'PATH pi should not be called\\n' >"\${PATH_PI_LOG}"\nexit 97`);
-
-	const result = spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-update.mjs"), "--extensions", "--agent-dir", agentDir, "--quiet"], {
-		cwd: repoRoot,
-		env: withNodeImport(scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: `${pathPiDir}:${process.env.PATH || ""}`,
-			PI_WRAPPER_LOG: piLog,
-			PATH_PI_LOG: pathPiLog,
-			TLH_TEST_RTK_LOG: rtkLog,
-		}), managedRtkSpawnPreload),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(result.status, 0, `Case 1 stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-	assert.equal(result.stdout, "");
-	const piRecord = Object.fromEntries(readFileSync(piLog, "utf8").trim().split(/\r?\n/).map((line) => {
-		const separator = line.indexOf("=");
-		return [line.slice(0, separator), line.slice(separator + 1)];
-	}));
-	// Absolute private runtime pi was used — not any PATH-based pi.
-	assert.equal(piRecord.cmd, join(runtimeBinDir, "pi"), "expected absolute private runtime pi");
-	assert.equal(piRecord.argv, "update --extensions");
-	assert.equal(piRecord.agent, agentDir);
-	assert.equal(existsSync(pathPiLog), false, "PATH pi must not be called");
-
-	// Case 2: private runtime pi absent, non-dry-run — must hard-fail with a clear error.
-	const root2 = makeTempDir();
-	const agentDir2 = join(root2, "agent");
-	const homeDir2 = join(root2, "home");
-	mkdirSync(agentDir2, { recursive: true });
-	mkdirSync(homeDir2, { recursive: true });
-	t.after(() => rmSync(root2, { recursive: true, force: true }));
-	// No private runtime pi created at root2/runtime/bin/pi.
-	const missingResult = spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-update.mjs"), "--extensions", "--agent-dir", agentDir2], {
-		cwd: repoRoot,
-		env: withNodeImport(scrubInstallerEnv({
-			HOME: homeDir2,
-			PATH: `${pathPiDir}:${process.env.PATH || ""}`,
-			TLH_TEST_RTK_LOG: rtkLog,
-		}), managedRtkSpawnPreload),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.notEqual(missingResult.status, 0, "expected hard failure when private runtime pi is missing");
-	assert.equal(missingResult.stdout, "");
-	assert.match(missingResult.stderr, /private runtime pi not found/);
-	assert.match(missingResult.stderr, /tlh update/);
-	// Dry-run must NOT hard-fail even when the binary is absent.
-	const dryRunResult = spawnSync(process.execPath, [join(repoRoot, "scripts/tlh-update.mjs"), "--extensions", "--dry-run", "--agent-dir", agentDir2], {
-		cwd: repoRoot,
-		env: withNodeImport(scrubInstallerEnv({
-			HOME: homeDir2,
-			PATH: process.env.PATH || "",
-			TLH_TEST_RTK_LOG: rtkLog,
-		}), managedRtkSpawnPreload),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(dryRunResult.status, 0, `dry-run must succeed even without runtime pi: ${dryRunResult.stderr}`);
-	assert.match(dryRunResult.stdout, /Would run:/);
-	assert.match(dryRunResult.stdout, new RegExp(join(root2, "runtime", "bin", "pi").replace(/[.*+?^${}()|[\]\\]/g, "\\$&")));
-	if (managedRtkSupportedTestPlatform) {
-		const rtkCalls = readJsonLines(rtkLog);
-		assert.equal(rtkCalls.length, 2);
-		assert.deepEqual(rtkCalls[0]?.args, [join(repoRoot, "scripts", "tlh-rtk.mjs"), "--agent-dir", agentDir, "--target", join(agentDir, "bin", "rtk"), "--quiet", "install-managed"]);
-		assert.deepEqual(rtkCalls[1]?.args, [join(repoRoot, "scripts", "tlh-rtk.mjs"), "--agent-dir", agentDir2, "--target", join(agentDir2, "bin", "rtk"), "--dry-run", "--detail", "install-managed"]);
-	} else {
-		assert.equal(existsSync(rtkLog), false);
-	}
-});
-
-
-
-test("wrapper plain update helper does not prepend executable --pi-cmd directory", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const pinnedPiDir = join(root, "pinned-pi");
-	const cwdDir = join(root, "cwd");
-	const updateLog = join(root, "update.json");
-	const nodeDir = dirname(process.execPath);
-	const bashPath = spawnSync("sh", ["-c", "command -v bash"], { encoding: "utf8" }).stdout.trim() || "/bin/bash";
-	const bashDir = dirname(bashPath);
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageRoot, { recursive: true });
-	mkdirSync(cwdDir, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(pinnedPiDir, { recursive: true });
-	writeFileSync(join(pinnedPiDir, "pi"), "#!/bin/sh\nprintf 'unexpected args: %s\\n' \"$*\" >&2\nexit 1\n", "utf8");
-	chmodSync(join(pinnedPiDir, "pi"), 0o755);
-	writeFileSync(join(agentDir, "tlh", "recover-update.mjs"), `import { writeFileSync } from "node:fs";\nwriteFileSync(process.env.TLH_UPDATE_LOG, JSON.stringify({ argv: process.argv.slice(2), env: { PI_CODING_AGENT_DIR: process.env.PI_CODING_AGENT_DIR, PATH: process.env.PATH } }));\n`, "utf8");
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--package-root",
-		packageRoot,
-		"--pi-cmd",
-		join(pinnedPiDir, "pi"),
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const result = spawnSync(wrapper, ["update", "--dry-run"], {
-		cwd: cwdDir,
-		env: scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: ["", ".", cwdDir, agentBin, nodeDir, bashDir].join(delimiter),
-			TLH_UPDATE_LOG: updateLog,
-		}),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-
-	assert.equal(result.status, 0, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-
-	const updateRecord = JSON.parse(readFileSync(updateLog, "utf8"));
-	assert.deepEqual(updateRecord.argv, [
-		"--agent-dir",
-		agentDir,
-		"--bin-dir",
-		binDir,
-		"--wrapper-name",
-		"tlh",
-		"--dry-run",
-	]);
-	assert.equal(updateRecord.env.PI_CODING_AGENT_DIR, agentDir);
-
-	const updatePathEntries = updateRecord.env.PATH.split(delimiter);
-	assert.equal(updatePathEntries[0], nodeDir, `expected sanitized PATH first; got ${updatePathEntries.join(delimiter)}`);
-	assert.equal(updatePathEntries.includes(pinnedPiDir), false, `did not expect pinned_dir in PATH: ${updatePathEntries.join(delimiter)}`);
-	assert.equal(updatePathEntries.includes(""), false);
-	assert.equal(updatePathEntries.includes("."), false);
-	assert.equal(updatePathEntries.includes(cwdDir), false);
-	assert.equal(updatePathEntries.includes(agentBin), false);
-});
-
-// ── NODE_COMPILE_CACHE wrapper tests ────────────────────────────────────────
-
-test("wrapper pi exec path exports NODE_COMPILE_CACHE pointing at the runtime prefix", (t) => {
-	const root = makeTempDir("tlh-wrapper-node-compile-cache-");
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const agentBin = join(agentDir, "bin");
-	const runtimeDir = join(root, "runtime");
-	const pinnedPiDir = join(runtimeDir, "bin");
-	const binDir = join(root, "bin");
-	const packageRoot = join(root, "package");
-	const piLog = join(root, "pi.txt");
-	mkdirSync(agentBin, { recursive: true });
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageRoot, { recursive: true });
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	// Fake pi logs NODE_COMPILE_CACHE so we can assert its value.
-	writeFakePi(pinnedPiDir, [
-		`{ printf 'compile_cache=%s\\n' "\${NODE_COMPILE_CACHE:-}"; } >"${piLog}"`,
-		"exit 0",
-	].join("\n"));
-
-	runHelper("scripts/tlh-wrapper.mjs", [
-		"--agent-dir", agentDir,
-		"--bin-dir", binDir,
-		"--wrapper-name", "tlh",
-		"--package-root", packageRoot,
-		"--pi-cmd", join(pinnedPiDir, "pi"),
-	], { homeDir });
-
-	const wrapper = join(binDir, "tlh");
-	const result = spawnSync(wrapper, ["chat"], {
-		env: scrubInstallerEnv({ HOME: homeDir, PATH: process.env.PATH || "" }),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	assert.equal(result.status, 0, `stdout:\n${result.stdout}\nstderr:\n${result.stderr}`);
-
-	const piRecord = Object.fromEntries(
-		readFileSync(piLog, "utf8").trim().split(/\r?\n/).map((line) => {
-			const separator = line.indexOf("=");
-			return [line.slice(0, separator), line.slice(separator + 1)];
-		}),
-	);
-
-	// NODE_COMPILE_CACHE must point to <runtime-prefix>/node-compile-cache.
-	// runtimeDir = dirname(pinnedPiDir), which is dirname(dirname(piCmd)).
-	assert.equal(
-		piRecord.compile_cache,
-		join(runtimeDir, "node-compile-cache"),
-		`NODE_COMPILE_CACHE must be runtimeDir/node-compile-cache; got: ${piRecord.compile_cache}`,
-	);
-});
-
-// Regression (tlht-5php): post-install validation — broken/wrong-version npm install must
-// throw, leaving any user-owned ~/.local/bin/pi untouched.
-test("stage-1 installPiIfNeeded: broken npm install (wrong pi version) throws", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const packageDir = join(root, "package-source");
-	const npmLog = join(root, "npm.log");
-	const templateDir = join(root, "pi-template");
-	const runtimeDir = join(root, "runtime");
-	const installedPiPath = join(runtimeDir, "bin", "pi");
-	const legacyBin = join(homeDir, ".local", "bin");
-	const legacyPiPath = join(legacyBin, "pi");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(legacyBin, { recursive: true });
-
-	// Install-state: piInstalledByTlh=true
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		track: "ref",
-		ref: "main",
-		packageSource: packageDir,
-		packageSourceIsDefault: false,
-		piInstalledByTlh: true,
-	}, null, 2));
-
-	// Legacy pi at ~/.local/bin/pi — must NOT be removed when install fails.
-	writeFakePi(legacyBin, `if [[ "\${1:-}" == "--version" ]]; then printf '${TLH_PINNED_PI_VERSION}\\n'; exit 0; fi\nexit 0`);
-
-	// Template pi with a clearly wrong non-pinned version (0.80.7) — simulates a broken npm install.
-	writeFakePi(templateDir, "if [[ \"${1:-}\" == \"--version\" ]]; then printf '0.80.7\\n'; exit 0; fi\nexit 0");
-
-	// Fake npm: always installs the wrong-version template pi.
-	writeFakeNpmInstaller(fakebin, {
-		npmLog,
-		templatePiPath: join(templateDir, "pi"),
-		installedPiPath,
-	});
-	writeFakeCommand(fakebin, "git", "exit 0");
-	writeFakeTk(fakebin);
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: safeInstallerPath(fakebin),
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-	});
-	const result = runInstaller([
-		"--agent-dir", agentDir,
-		"--bin-dir", binDir,
-		"--no-settings",
-		"--no-wrapper",
-		"--pi-installed-by-tlh", "true",
-	], env);
-	const output = `${result.stdout}\n${result.stderr}`;
-
-	// Installer must fail: freshly installed pi has wrong version.
-	assert.notEqual(result.status, 0, `expected installer to fail on wrong-version pi, got exit 0:\n${output}`);
-	assert.match(output, /0\.80\.7/, "error output should mention the wrong version");
-
-	// ~/.local/bin/pi must NOT have been removed (install threw before any cleanup could run).
-	assert.equal(existsSync(legacyPiPath), true, "user-owned ~/.local/bin/pi was removed despite installer throwing");
-
-	// npm uninstall must NOT have been called.
-	const npmLinesC = existsSync(npmLog)
-		? readFileSync(npmLog, "utf8").trim().split(/\r?\n/).filter(Boolean)
-		: [];
-	assert.equal(
-		npmLinesC.filter((line) => line.startsWith("uninstall ")).length,
-		0,
-		`npm uninstall must not have been called; npm log: ${npmLinesC.join(", ")}`,
-	);
-});
-
-// Regression (tlht-5php, blocker): piInstalledByTlh=true, private runtime ABSENT at start,
-// user-owned ~/.local/bin/pi@0.80.6 present.  The installer must provision the private
-// runtime and succeed — without removing or executing ~/.local/bin/pi.
-test("stage-1 regression (tlht-5php): installer never removes or execs user-owned ~/.local/bin/pi when piInstalledByTlh=true and private runtime is absent", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const packageDir = join(root, "package-source");
-	const npmLog = join(root, "npm.log");
-	const templateDir = join(root, "pi-template");
-	const runtimeDir = join(root, "runtime");
-	const installedPiPath = join(runtimeDir, "bin", "pi");
-	const piLog = join(root, "pi.log");
-	const legacyBin = join(homeDir, ".local", "bin");
-	const legacyPiPath = join(legacyBin, "pi");
-	const legacyPiInvocationLog = join(root, "legacy-pi.log");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(legacyBin, { recursive: true });
-
-	// Install-state: piInstalledByTlh=true (the blocker scenario: set by old model or stale state).
-	// No private runtime exists at start — runtimeDir is absent.
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		track: "ref",
-		ref: "main",
-		packageSource: packageDir,
-		packageSourceIsDefault: false,
-		piInstalledByTlh: true,
-	}, null, 2));
-
-	// User-owned ~/.local/bin/pi@0.80.6 — must NOT be removed or invoked by the installer.
-	writeFakePi(legacyBin, [
-		`printf '%s\\n' "$*" >>"${legacyPiInvocationLog}"`,
-		`if [[ "\${1:-}" == "--version" ]]; then printf '${TLH_PINNED_PI_VERSION}\\n'; exit 0; fi`,
-		"exit 0",
-	].join("\n"));
-
-	// Template pi for npm to install as the new private runtime (correct pinned version).
-	writeLoggingPi(templateDir, piLog, TLH_PINNED_PI_VERSION);
-
-	// Fake npm: logs all invocations, handles install only (copies template to runtime path).
-	// Any npm uninstall call would indicate the installer is (incorrectly) trying to remove ~/.local.
-	writeFakeCommand(fakebin, "npm", [
-		`printf '%s\\n' "$*" >>"${npmLog}"`,
-		`case "$1" in`,
-		`  install)`,
-		`    mkdir -p "${join(runtimeDir, "bin")}"`,
-		`    cp "${join(templateDir, "pi")}" "${installedPiPath}"`,
-		`    chmod +x "${installedPiPath}"`,
-		`    ;;`,
-		`esac`,
-	].join("\n"));
-	writeFakeCommand(fakebin, "git", "exit 0");
-	writeFakeTk(fakebin);
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: safeInstallerPath(fakebin),
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-	});
-	const result = runInstaller([
-		"--agent-dir", agentDir,
-		"--bin-dir", binDir,
-		"--no-settings",
-		"--no-wrapper",
-		"--pi-installed-by-tlh", "true",
-	], env);
-	const output = `${result.stdout}\n${result.stderr}`;
-
-	// Installer must succeed — the private runtime was freshly installed.
-	assert.equal(result.status, 0, `installer failed unexpectedly:\n${output}`);
-	assert.equal(existsSync(installedPiPath), true, "private runtime was not installed");
-
-	// User-owned ~/.local/bin/pi must still be present (not removed).
-	assert.equal(existsSync(legacyPiPath), true, "user-owned ~/.local/bin/pi was incorrectly removed by the installer");
-
-	// npm uninstall must NOT have been called for the ~/.local prefix.
-	const npmCalls = existsSync(npmLog)
-		? readFileSync(npmLog, "utf8").trim().split(/\r?\n/).filter(Boolean)
-		: [];
-	const uninstallCalls = npmCalls.filter((line) => line.startsWith("uninstall "));
-	assert.equal(
-		uninstallCalls.length,
-		0,
-		`npm uninstall must never be called by the installer; got: ${uninstallCalls.join(", ")}`,
-	);
-
-	// ~/.local/bin/pi must NOT have been executed by the installer.
-	assert.equal(
-		existsSync(legacyPiInvocationLog),
-		false,
-		"user-owned ~/.local/bin/pi was invoked by the installer (should never happen)",
-	);
-});
-
-// Scenario (d): uninstall.sh — piInstalledByTlh=true, no private runtime, user-owned
-// legacy ~/.local/bin/pi.  Without --force-include-pi the pi is never removed
-// (hint printed).  With --force-include-pi it is removed via npm.
-test("uninstall.sh does not remove legacy ~/.local/bin/pi without --force-include-pi (prints hint); removes it with --force-include-pi", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const legacyBin = join(homeDir, ".local", "bin");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakeNpmDir = join(root, "fakenpm");
-	const npmLog = join(root, "npm-uninstall.log");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	// Setup: agent dir with TLH ownership marker and piInstalledByTlh=true.
-	// NO private runtime (runtimeDir does not exist).
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(legacyBin, { recursive: true });
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		piInstalledByTlh: true,
-	}, null, 2));
-
-	// User-owned legacy pi at ~/.local/bin/pi.
-	writeFileSync(join(legacyBin, "pi"), "#!/bin/sh\nprintf '0.80.6\\n'\n", "utf8");
-	chmodSync(join(legacyBin, "pi"), 0o755);
-
-	// ── (d1) dry-run without --force-include-pi: hint printed, pi NOT removed ──
-	const dryResult = spawnSync("bash", [join(repoRoot, "uninstall.sh"), "--dry-run", "--agent-dir", agentDir, "--bin-dir", binDir], {
-		cwd: repoRoot,
-		env: scrubInstallerEnv({ HOME: homeDir }),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	const dryOutput = `${dryResult.stdout}\n${dryResult.stderr}`;
-
-	assert.equal(dryResult.status, 0, `dry-run failed:\n${dryOutput}`);
-	assert.match(dryOutput, /force-include-pi/, "dry-run output must mention --force-include-pi hint");
-	assert.equal(existsSync(join(legacyBin, "pi")), true, "dry-run must not remove legacy pi");
-
-	// ── (d2) with --force-include-pi: npm uninstall called, pi removed ─────────
-	// Fake npm: log the call and remove the legacy pi binary.
-	mkdirSync(fakeNpmDir, { recursive: true });
-	writeFileSync(join(fakeNpmDir, "npm"), [
-		"#!/usr/bin/env bash",
-		"set -euo pipefail",
-		`printf '%s\\n' "$*" >>"${npmLog}"`,
-		`if [[ "$1" == "uninstall" ]]; then`,
-		`  rm -f "${join(legacyBin, "pi")}"`,
-		`fi`,
-	].join("\n"), "utf8");
-	chmodSync(join(fakeNpmDir, "npm"), 0o755);
-
-	// Re-create the agent dir with marker (the dry-run left it intact; we need it for the real run).
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		piInstalledByTlh: true,
-	}, null, 2));
-
-	const forceResult = spawnSync("bash", [join(repoRoot, "uninstall.sh"), "--force-include-pi", "--agent-dir", agentDir, "--bin-dir", binDir], {
-		cwd: repoRoot,
-		env: scrubInstallerEnv({ HOME: homeDir, PATH: `${fakeNpmDir}:${process.env.PATH || ""}` }),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	const forceOutput = `${forceResult.stdout}\n${forceResult.stderr}`;
-
-	assert.equal(forceResult.status, 0, `force-include-pi uninstall failed:\n${forceOutput}`);
-	// npm must have been called for uninstall.
-	assert.equal(existsSync(npmLog), true, "npm uninstall was not called with --force-include-pi");
-	const npmCalls = readFileSync(npmLog, "utf8").trim().split(/\r?\n/).filter(Boolean);
-	assert.ok(npmCalls.some((line) => line.startsWith("uninstall ")), `expected npm uninstall; got: ${npmCalls.join(", ")}`);
-	// Legacy pi must have been removed.
-	assert.equal(existsSync(join(legacyBin, "pi")), false, "legacy pi should have been removed with --force-include-pi");
-});
-
-test("uninstall.sh regression (tlht-h7vq): migrated runtime marker preserves nested foreign packages during surgical uninstall", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const profileRoot = join(root, "profile");
-	const agentDir = join(profileRoot, "agent");
-	const runtimeDir = join(profileRoot, "runtime");
-	const binDir = join(root, "bin");
-	const fakeNpmDir = join(root, "fakenpm");
-	const npmLog = join(root, "npm.log");
-	const markerPath = join(runtimeDir, ".tlh-runtime-owned");
-	const tlhPackageDir = join(runtimeDir, "lib", "node_modules", "@earendil-works", "pi-coding-agent");
-	const foreignPackageDir = join(runtimeDir, "lib", "node_modules", "foreign-package");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(join(runtimeDir, "bin"), { recursive: true });
-	mkdirSync(tlhPackageDir, { recursive: true });
-	mkdirSync(foreignPackageDir, { recursive: true });
-	mkdirSync(binDir, { recursive: true });
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		piInstalledByTlh: true,
-	}, null, 2));
-	writeFileSync(join(runtimeDir, "bin", "pi"), "#!/bin/sh\n", "utf8");
-	chmodSync(join(runtimeDir, "bin", "pi"), 0o755);
-	writeFileSync(markerPath, JSON.stringify({
-		schemaVersion: 1,
-		packageName: "@earendil-works/pi-coding-agent",
-		runtimeAbsPath: realpathSync(runtimeDir),
-		origin: "migrated",
-	}), "utf8");
-	writeFileSync(join(tlhPackageDir, "package.json"), JSON.stringify({ name: "@earendil-works/pi-coding-agent" }, null, 2));
-	writeFileSync(join(foreignPackageDir, "package.json"), JSON.stringify({ name: "foreign-package" }, null, 2));
-
-	const dryResult = spawnSync("bash", [join(repoRoot, "uninstall.sh"), "--dry-run", "--agent-dir", agentDir, "--bin-dir", binDir], {
-		cwd: repoRoot,
-		env: scrubInstallerEnv({ HOME: homeDir }),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	const dryOutput = `${dryResult.stdout}\n${dryResult.stderr}`;
-
-	assert.equal(dryResult.status, 0, `dry-run failed:\n${dryOutput}`);
-	assert.match(
-		dryOutput,
-		new RegExp(`would remove migrated TLH pi from shared runtime \\(npm\\): npm uninstall -g --ignore-scripts --prefix "${escapeRegExp(runtimeDir)}" @earendil-works/pi-coding-agent`),
-	);
-	assert.doesNotMatch(dryOutput, new RegExp(`would remove private runtime: rm -rf ${escapeRegExp(runtimeDir)}`));
-	assert.equal(existsSync(foreignPackageDir), true, "dry-run must not remove nested foreign package");
-
-	writeFakeCommand(fakeNpmDir, "npm", [
-		`printf '%s\\n' "$*" >>"${npmLog}"`,
-		`if [[ "$1" != "uninstall" ]]; then printf 'unexpected npm command: %s\\n' "$*" >&2; exit 98; fi`,
-		`rm -f "${join(runtimeDir, "bin", "pi")}"`,
-		`rm -rf "${tlhPackageDir}"`,
-	].join("\n"));
-
-	const realResult = spawnSync("bash", [join(repoRoot, "uninstall.sh"), "--agent-dir", agentDir, "--bin-dir", binDir], {
-		cwd: repoRoot,
-		env: scrubInstallerEnv({ HOME: homeDir, PATH: `${fakeNpmDir}:${process.env.PATH || ""}` }),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	const realOutput = `${realResult.stdout}\n${realResult.stderr}`;
-
-	assert.equal(realResult.status, 0, `real uninstall failed:\n${realOutput}`);
-	assert.deepEqual(readFileSync(npmLog, "utf8").trim().split(/\r?\n/).filter(Boolean), [
-		`uninstall -g --ignore-scripts --prefix ${runtimeDir} @earendil-works/pi-coding-agent`,
-	]);
-	assert.equal(existsSync(agentDir), false, "agent dir should be removed after uninstall");
-	assert.equal(existsSync(join(runtimeDir, "bin", "pi")), false, "TLH runtime launcher should be removed surgically");
-	assert.equal(existsSync(tlhPackageDir), false, "TLH runtime package should be removed surgically");
-	assert.equal(existsSync(markerPath), false, "migrated runtime ownership marker should be cleared after uninstall");
-	assert.equal(existsSync(foreignPackageDir), true, "nested foreign package must survive migrated runtime uninstall");
-	assert.equal(readFileSync(join(foreignPackageDir, "package.json"), "utf8"), JSON.stringify({ name: "foreign-package" }, null, 2));
-	assert.equal(existsSync(runtimeDir), true, "shared runtime prefix must survive migrated runtime uninstall");
-});
-
-test("stage-1 refuses to install into a runtime prefix containing a foreign top-level entry", (t) => {
-	const scenarios = [
-		{ label: "normal file", foreignName: "userdata.txt", dryRun: false },
-		{ label: "dotfile", foreignName: "..userdata", dryRun: false },
-		{ label: "normal file (dry-run)", foreignName: "userdata.txt", dryRun: true },
-	];
-
-	for (const scenario of scenarios) {
-		const root = makeTempDir(`tlh-install-runtime-guard-${scenario.label.replace(/[^a-z0-9]+/g, "-")}-`);
-		const homeDir = join(root, "home");
-		const agentDir = join(root, "agent");
-		const binDir = join(root, "bin");
-		const fakebin = join(root, "fakebin");
-		const packageDir = join(root, "package-source");
-		const templateDir = join(root, "pi-template");
-		const npmLog = join(root, "npm.log");
-		const piLog = join(root, "pi.log");
-		const runtimeDir = join(root, "runtime"); // sibling of agentDir per runtimePrefix(config)
-		t.after(() => rmSync(root, { recursive: true, force: true }));
-
-		mkdirSync(homeDir, { recursive: true });
-		mkdirSync(packageDir, { recursive: true });
-		// Pre-seed the runtime prefix with a foreign top-level entry.
-		mkdirSync(runtimeDir, { recursive: true });
-		const foreignPath = join(runtimeDir, scenario.foreignName);
-		writeFileSync(foreignPath, "user content\n");
-
-		writeFakeTk(fakebin);
-		writeLoggingPi(templateDir, piLog, TLH_PINNED_PI_VERSION);
-		writeFakeNpmInstaller(fakebin, {
-			npmLog,
-			templatePiPath: join(templateDir, "pi"),
-			installedPiPath: join(runtimeDir, "bin", "pi"),
-		});
-
-		const env = scrubInstallerEnv({
-			HOME: homeDir,
-			PATH: `${fakebin}:${process.env.PATH || ""}`,
-			TLH_PACKAGE_SOURCE: packageDir,
-			TLH_SKIP_GNOSIS_INSTALL: "1",
-		});
-		const args = ["--agent-dir", agentDir, "--bin-dir", binDir, "--no-wrapper", "--no-settings"];
-		if (scenario.dryRun) args.unshift("--dry-run");
-		const result = runInstaller(args, env);
-		const output = `${result.stdout}\n${result.stderr}`;
-
-		// Must fail — ownership guard surfaces even in dry-run.
-		assert.notEqual(result.status, 0, `${scenario.label}: expected failure but installer succeeded\n${output}`);
-		// Error message must be actionable and name the ownership marker.
-		assert.match(output, /is not TLH-owned/, `${scenario.label}: error must name the ownership problem`);
-		assert.match(output, /--agent-dir/, `${scenario.label}: error must mention --agent-dir`);
-		// Foreign file must be left untouched.
-		assert.equal(existsSync(foreignPath), true, `${scenario.label}: foreign file must still exist after refused install`);
-		assert.equal(readFileSync(foreignPath, "utf8"), "user content\n", `${scenario.label}: foreign file content must be unchanged`);
-		// npm must not have been invoked.
-		assert.equal(existsSync(npmLog), false, `${scenario.label}: npm must not be invoked when guard refuses`);
-	}
-});
-
-test("stage-1 installs normally when runtime prefix exists but is empty", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const packageDir = join(root, "package-source");
-	const templateDir = join(root, "pi-template");
-	const npmLog = join(root, "npm.log");
-	const piLog = join(root, "pi.log");
-	const runtimeDir = join(root, "runtime");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	// Pre-create the runtime prefix as an empty directory.
-	mkdirSync(runtimeDir, { recursive: true });
-
-	writeFakeTk(fakebin);
-	writeLoggingPi(templateDir, piLog, TLH_PINNED_PI_VERSION);
-	writeFakeNpmInstaller(fakebin, {
-		npmLog,
-		templatePiPath: join(templateDir, "pi"),
-		installedPiPath: join(runtimeDir, "bin", "pi"),
-	});
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: `${fakebin}:${process.env.PATH || ""}`,
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-	});
-	const result = runInstaller(
-		["--agent-dir", agentDir, "--bin-dir", binDir, "--no-wrapper", "--no-settings"],
-		env,
-	);
-	const output = `${result.stdout}\n${result.stderr}`;
-
-	assert.equal(result.status, 0, `install into empty runtime prefix failed:\n${output}`);
-	// npm must have been called to install pi.
-	assert.equal(existsSync(npmLog), true, "npm was not called for empty runtime prefix");
-});
-
-// ---------------------------------------------------------------------------
-// Runtime ownership marker tests (tlht-7mx4)
-// ---------------------------------------------------------------------------
-
-test("RUNTIME_MARKER_FILENAME .tlh-runtime-owned is in RUNTIME_OWNED_TOPLEVEL", () => {
-	assert.equal(RUNTIME_MARKER_FILENAME, ".tlh-runtime-owned");
-	assert.ok(RUNTIME_OWNED_TOPLEVEL.has(RUNTIME_MARKER_FILENAME), "marker filename must be in RUNTIME_OWNED_TOPLEVEL allow-list");
-});
-
-test("runtime ownership: pristine/absent prefix is accepted and marker origin=created is written", (t) => {
-	// runStage1LocalPackageInstall starts with no pre-existing runtime prefix.
-	const { result, agentDir } = runStage1LocalPackageInstall(t, { noSettings: true });
-	const output = `${result.stdout}\n${result.stderr}`;
-	const runtimeDir = join(dirname(agentDir), "runtime");
-	const markerPath = join(runtimeDir, ".tlh-runtime-owned");
-
-	assert.equal(result.status, 0, output);
-	assert.ok(existsSync(markerPath), "ownership marker must be written after fresh install");
-	const marker = JSON.parse(readFileSync(markerPath, "utf8"));
-	assert.equal(marker.schemaVersion, 1);
-	assert.equal(marker.packageName, "@earendil-works/pi-coding-agent");
-	assert.equal(marker.origin, "created");
-	assert.equal(typeof marker.runtimeAbsPath, "string");
-	assert.ok(marker.runtimeAbsPath.length > 0);
-});
-
-test("runtime ownership: non-empty unmarked prefix without provenance is refused", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const runtimeDir = join(root, "runtime"); // sibling of agentDir = runtimePrefix(config)
-	const packageDir = join(root, "package-source");
-	const npmLog = join(root, "npm.log");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	// Pre-create a non-empty runtime prefix WITHOUT the ownership marker
-	// and WITHOUT an install-state (no piInstalledByTlh=true provenance).
-	mkdirSync(join(runtimeDir, "bin"), { recursive: true });
-	writeFileSync(join(runtimeDir, "bin", "some-binary"), "#!/bin/sh\nexit 0\n", "utf8");
-	writeFakeCommand(fakebin, "git", "exit 0");
-	writeFakeTk(fakebin);
-	writeFakeCommand(fakebin, "npm", `printf '%s\\n' "$*" >>"${npmLog}"\nexit 97`);
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: safeInstallerPath(fakebin),
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-	});
-	const result = runInstaller([
-		"--agent-dir", agentDir,
-		"--bin-dir", binDir,
-		"--no-settings",
-		"--no-wrapper",
-	], env);
-	const output = `${result.stdout}\n${result.stderr}`;
-
-	assert.notEqual(result.status, 0, `expected failure but installer succeeded\n${output}`);
-	assert.match(output, /is not TLH-owned/, "error must name the ownership problem");
-	assert.match(output, /\.tlh-runtime-owned/, "error must mention the marker filename");
-	assert.match(output, /--agent-dir/, "error must mention --agent-dir");
-	// npm must not be invoked when ownership is refused.
-	assert.equal(existsSync(npmLog), false, "npm must not be called when ownership guard refuses");
-	// No marker must be written.
-	assert.equal(existsSync(join(runtimeDir, ".tlh-runtime-owned")), false, "marker must not be written on refusal");
-});
-
-test("runtime ownership: existing valid marker (path-matched) is accepted on reuse", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const runtimeDir = join(root, "runtime");
-	const packageDir = join(root, "package-source");
-	const npmLog = join(root, "npm.log");
-	const piLog = join(root, "pi.log");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	mkdirSync(join(runtimeDir, "bin"), { recursive: true });
-	writeFakeCommand(fakebin, "git", "exit 0");
-	writeFakeTk(fakebin);
-	// npm must NOT be called (valid pi exists).
-	writeFakeCommand(fakebin, "npm", `printf '%s\\n' "$*" >>"${npmLog}"\nexit 97`);
-	writeLoggingPi(join(runtimeDir, "bin"), piLog, TLH_PINNED_PI_VERSION);
-
-	// Write a valid ownership marker into the pre-existing runtime.
-	const realRuntimeDir = realpathSync(runtimeDir);
-	writeFileSync(join(runtimeDir, ".tlh-runtime-owned"), JSON.stringify({
-		schemaVersion: 1,
-		packageName: "@earendil-works/pi-coding-agent",
-		runtimeAbsPath: realRuntimeDir,
-		origin: "created",
-	}), "utf8");
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: safeInstallerPath(fakebin),
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-	});
-	const result = runInstaller([
-		"--agent-dir", agentDir,
-		"--bin-dir", binDir,
-		"--no-settings",
-		"--no-wrapper",
-	], env);
-	const output = `${result.stdout}\n${result.stderr}`;
-
-	assert.equal(result.status, 0, output);
-	// npm must not be called (valid pi + valid marker = reuse path).
-	assert.equal(existsSync(npmLog), false, "npm must not be called when valid marker present and pi is valid");
-	// pi --version was called (reuse path version check).
-	const piRecords = readPiLogRecords(piLog);
-	assert.ok(piRecords.some((r) => r.command === "--version"), "pi --version must be called on reuse");
-	// Marker must be refreshed (still present) after reuse.
-	const marker = JSON.parse(readFileSync(join(runtimeDir, ".tlh-runtime-owned"), "utf8"));
-	assert.equal(marker.origin, "created", "marker origin must be preserved on refresh");
-});
-
-test("runtime ownership: non-empty unmarked prefix with piInstalledByTlh=true is migrated", (t) => {
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const runtimeDir = join(root, "runtime");
-	const packageDir = join(root, "package-source");
-	const npmLog = join(root, "npm.log");
-	const piLog = join(root, "pi.log");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	mkdirSync(join(runtimeDir, "bin"), { recursive: true });
-	writeFakeCommand(fakebin, "git", "exit 0");
-	writeFakeTk(fakebin);
-	// npm must NOT be called (valid pi exists, migration only writes marker).
-	writeFakeCommand(fakebin, "npm", `printf '%s\\n' "$*" >>"${npmLog}"\nexit 97`);
-	writeLoggingPi(join(runtimeDir, "bin"), piLog, TLH_PINNED_PI_VERSION);
-
-	// Install-state carries piInstalledByTlh=true (provenance from a prior install).
-	writeFileSync(join(agentDir, "tlh", "install-state.json"), JSON.stringify({
-		schemaVersion: 1,
-		repo: "diegopetrucci/the-last-harness",
-		track: "ref",
-		ref: "main",
-		packageSource: packageDir,
-		packageSourceIsDefault: false,
-		piInstalledByTlh: true,
-	}, null, 2));
-	// No marker file exists yet.
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: safeInstallerPath(fakebin),
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-	});
-	const result = runInstaller([
-		"--agent-dir", agentDir,
-		"--bin-dir", binDir,
-		"--no-settings",
-		"--no-wrapper",
-	], env);
-	const output = `${result.stdout}\n${result.stderr}`;
-
-	assert.equal(result.status, 0, output);
-	// npm must not be called (valid pi + migration path).
-	assert.equal(existsSync(npmLog), false, "npm must not be called during provenance-gated migration");
-	// Marker must be written with origin='migrated'.
-	const markerPath = join(runtimeDir, ".tlh-runtime-owned");
-	assert.ok(existsSync(markerPath), "ownership marker must be written after migration");
-	const marker = JSON.parse(readFileSync(markerPath, "utf8"));
-	assert.equal(marker.origin, "migrated", "migrated marker must carry origin=migrated");
-	assert.equal(marker.schemaVersion, 1);
-	assert.equal(marker.packageName, "@earendil-works/pi-coding-agent");
-});
-
-test("runtime ownership: symlinked runtime prefix is refused", (t) => {
-	if (process.platform === "win32") {
-		return; // symlinks not reliably available on Windows
-	}
-	const root = makeTempDir();
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const fakebin = join(root, "fakebin");
-	const runtimeTarget = join(root, "runtime-real"); // the real directory
-	const runtimeLink = join(root, "runtime"); // sibling of agentDir; symlink
-	const packageDir = join(root, "package-source");
-	const npmLog = join(root, "npm.log");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	mkdirSync(packageDir, { recursive: true });
-	mkdirSync(agentDir, { recursive: true });
-	mkdirSync(runtimeTarget, { recursive: true });
-	// Create the runtime dir as a symlink to a real directory.
-	symlinkSync(runtimeTarget, runtimeLink, "dir");
-	writeFakeCommand(fakebin, "git", "exit 0");
-	writeFakeTk(fakebin);
-	writeFakeCommand(fakebin, "npm", `printf '%s\\n' "$*" >>"${npmLog}"\nexit 97`);
-
-	const env = scrubInstallerEnv({
-		HOME: homeDir,
-		PATH: safeInstallerPath(fakebin),
-		TLH_PACKAGE_SOURCE: packageDir,
-		TLH_SKIP_GNOSIS_INSTALL: "1",
-	});
-	const result = runInstaller([
-		"--agent-dir", agentDir,
-		"--bin-dir", binDir,
-		"--no-settings",
-		"--no-wrapper",
-	], env);
-	const output = `${result.stdout}\n${result.stderr}`;
-
-	assert.notEqual(result.status, 0, `expected failure on symlinked runtime but installer succeeded\n${output}`);
-	assert.match(output, /symlink/i, "error must mention symlink");
-	assert.match(output, /runtime/i, "error must mention runtime");
-	// npm must not be invoked.
-	assert.equal(existsSync(npmLog), false, "npm must not be called when symlink guard refuses");
-});
-
-// ---------------------------------------------------------------------------
-// Uninstall ownership-marker gate tests (tlht-7son)
-// ---------------------------------------------------------------------------
-
-// Helper: write agentDir install-state.json with piInstalledByTlh=true
-function writeAgentInstallState(agentDir, extra = {}) {
-	mkdirSync(join(agentDir, "tlh"), { recursive: true });
-	writeFileSync(
-		join(agentDir, "tlh", "install-state.json"),
-		JSON.stringify({ schemaVersion: 1, repo: "diegopetrucci/the-last-harness", piInstalledByTlh: true, ...extra }),
-	);
-}
-
-// Helper: write a valid .tlh-runtime-owned marker into runtimeDir
-function writeRuntimeMarker(runtimeDir, overrides = {}) {
-	const realRuntimeDir = realpathSync(runtimeDir);
-	writeFileSync(
-		join(runtimeDir, ".tlh-runtime-owned"),
-		JSON.stringify({
-			schemaVersion: 1,
-			packageName: "@earendil-works/pi-coding-agent",
-			runtimeAbsPath: realRuntimeDir,
-			origin: "created",
-			...overrides,
-		}),
-	);
-}
-
-// Helper: create minimum valid TLH runtime layout (bin/pi + lib/node_modules/PI)
-function writeRuntimeLayout(runtimeDir) {
-	mkdirSync(join(runtimeDir, "bin"), { recursive: true });
-	mkdirSync(join(runtimeDir, "lib", "node_modules", "@earendil-works", "pi-coding-agent"), { recursive: true });
-	writeFileSync(join(runtimeDir, "bin", "pi"), "#!/bin/sh\nexit 0\n", "utf8");
-	chmodSync(join(runtimeDir, "bin", "pi"), 0o755);
-}
-
-// Helper: run uninstall.sh and return result + combined output
-function runUninstall(args, homeDir) {
-	const result = spawnSync("bash", [join(repoRoot, "uninstall.sh"), ...args], {
-		cwd: repoRoot,
-		env: scrubInstallerEnv({ HOME: homeDir }),
-		encoding: "utf8",
-		stdio: ["ignore", "pipe", "pipe"],
-	});
-	return { result, output: `${result.stdout}\n${result.stderr}` };
-}
-
-// (a) Marked TLH runtime (valid marker, path match, layout present, not symlink)
-// → dry-run plan includes rm -rf; real run removes the runtime dir.
-test("uninstall.sh: marked TLH runtime is planned and removed (valid marker + layout)", (t) => {
-	const root = makeTempDir("tlh-uninstall-marker-a-");
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const runtimeDir = join(root, "runtime"); // PROFILE_ROOT/runtime = dirname(agentDir)/runtime
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	writeAgentInstallState(agentDir);
-	writeRuntimeLayout(runtimeDir);
-	writeRuntimeMarker(runtimeDir);
-
-	// ── dry-run: plan must include rm -rf ──
-	const { result: dryResult, output: dryOutput } = runUninstall(
-		["--dry-run", "--agent-dir", agentDir, "--bin-dir", binDir],
-		homeDir,
-	);
-	assert.equal(dryResult.status, 0, `dry-run failed:\n${dryOutput}`);
-	assert.match(dryOutput, /rm -rf/, "dry-run must include rm -rf for marked runtime");
-	// dry-run must not actually remove anything
-	assert.equal(existsSync(runtimeDir), true, "dry-run must not remove runtime dir");
-
-	// ── real run: runtime dir is removed ──
-	const { result, output } = runUninstall(
-		["--agent-dir", agentDir, "--bin-dir", binDir],
-		homeDir,
-	);
-	assert.equal(result.status, 0, `uninstall failed:\n${output}`);
-	assert.equal(existsSync(runtimeDir), false, "runtime dir must be removed for marked TLH runtime");
-});
-
-// (b) Shared/unmarked prefix containing the pi package AND an unrelated package
-// → SKIP with conditional manual-removal hint; foreign package untouched.
-test("uninstall.sh: shared/unmarked runtime prefix is SKIPPED with conditional hint; foreign package preserved", (t) => {
-	const root = makeTempDir("tlh-uninstall-marker-b-");
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const runtimeDir = join(root, "runtime");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	writeAgentInstallState(agentDir);
-	// Layout present (pi package) + foreign package in lib/node_modules — NO marker.
-	writeRuntimeLayout(runtimeDir);
-	mkdirSync(join(runtimeDir, "lib", "node_modules", "some-other-tool"), { recursive: true });
-	writeFileSync(join(runtimeDir, "lib", "node_modules", "some-other-tool", "index.js"), "// user content\n", "utf8");
-
-	const { result, output } = runUninstall(
-		["--dry-run", "--agent-dir", agentDir, "--bin-dir", binDir],
-		homeDir,
-	);
-	assert.equal(result.status, 0, `uninstall failed:\n${output}`);
-	// Must NOT plan rm -rf (no marker)
-	assert.doesNotMatch(output, /would remove private runtime/, "must not plan runtime removal without marker");
-	// Must emit the conditional manual-removal hint, not a blind rm suggestion
-	assert.match(output, /leave it/i, "output must include conditional 'leave it' hint");
-	// Foreign package must still exist
-	assert.equal(
-		existsSync(join(runtimeDir, "lib", "node_modules", "some-other-tool", "index.js")),
-		true,
-		"foreign package must be preserved when runtime is skipped",
-	);
-});
-
-// (c) Malformed/missing marker → SKIP (fail-closed)
-test("uninstall.sh: malformed or missing marker causes SKIP (fail-closed)", (t) => {
-	const scenarios = [
-		{ label: "missing marker file", markerContent: null },
-		{ label: "empty marker file", markerContent: "" },
-		{ label: "non-JSON content", markerContent: "not json at all" },
-		{ label: "wrong schemaVersion", markerContent: JSON.stringify({ schemaVersion: 99, packageName: "@earendil-works/pi-coding-agent", runtimeAbsPath: "/tmp", origin: "created" }) },
-		{ label: "wrong packageName", markerContent: JSON.stringify({ schemaVersion: 1, packageName: "@other/package", runtimeAbsPath: "/tmp", origin: "created" }) },
-		{ label: "invalid origin", markerContent: JSON.stringify({ schemaVersion: 1, packageName: "@earendil-works/pi-coding-agent", runtimeAbsPath: "/tmp", origin: "alien" }) },
-	];
-
-	for (const scenario of scenarios) {
-		const root = makeTempDir(`tlh-uninstall-marker-c-${scenario.label.replace(/[^a-z0-9]+/g, "-")}-`);
-		const homeDir = join(root, "home");
-		const agentDir = join(root, "agent");
-		const binDir = join(root, "bin");
-		const runtimeDir = join(root, "runtime");
-		t.after(() => rmSync(root, { recursive: true, force: true }));
-
-		mkdirSync(homeDir, { recursive: true });
-		writeAgentInstallState(agentDir);
-		writeRuntimeLayout(runtimeDir);
-		if (scenario.markerContent !== null) {
-			writeFileSync(join(runtimeDir, ".tlh-runtime-owned"), scenario.markerContent, "utf8");
-		}
-
-		const { result, output } = runUninstall(
-			["--dry-run", "--agent-dir", agentDir, "--bin-dir", binDir],
-			homeDir,
-		);
-		// Must not plan removal
-		assert.equal(result.status, 0, `${scenario.label}: uninstall failed:\n${output}`);
-		assert.doesNotMatch(output, /would remove private runtime/, `${scenario.label}: must not plan rm -rf when marker is malformed/missing`);
-		// Must emit conditional hint
-		assert.match(output, /leave it/i, `${scenario.label}: must include conditional hint on skip`);
-	}
-});
-
-// (d) Marker runtimeAbsPath mismatch → SKIP
-test("uninstall.sh: runtimeAbsPath mismatch in marker causes SKIP", (t) => {
-	const root = makeTempDir("tlh-uninstall-marker-d-");
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const runtimeDir = join(root, "runtime");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	writeAgentInstallState(agentDir);
-	writeRuntimeLayout(runtimeDir);
-	// Write marker pointing to a *different* path
-	writeFileSync(
-		join(runtimeDir, ".tlh-runtime-owned"),
-		JSON.stringify({
-			schemaVersion: 1,
-			packageName: "@earendil-works/pi-coding-agent",
-			runtimeAbsPath: "/some/other/completely/different/path",
-			origin: "created",
-		}),
-		"utf8",
-	);
-
-	const { result, output } = runUninstall(
-		["--dry-run", "--agent-dir", agentDir, "--bin-dir", binDir],
-		homeDir,
-	);
-	assert.equal(result.status, 0, `uninstall failed:\n${output}`);
-	// Must NOT plan removal
-	assert.doesNotMatch(output, /would remove private runtime/, "must not plan rm -rf when runtimeAbsPath mismatches");
-	// Must emit the conditional hint
-	assert.match(output, /leave it/i, "must include conditional hint on path mismatch");
-});
-
-// (e) Marked runtime with .tlh-runtime-owned in allow-list: exclusivity check does not trip on marker dotfile.
-test("uninstall.sh: marker dotfile is in exclusivity allow-list; runtime with bin+lib+node-compile-cache+marker is planned for removal", (t) => {
-	const root = makeTempDir("tlh-uninstall-marker-e-");
-	const homeDir = join(root, "home");
-	const agentDir = join(root, "agent");
-	const binDir = join(root, "bin");
-	const runtimeDir = join(root, "runtime");
-	t.after(() => rmSync(root, { recursive: true, force: true }));
-
-	mkdirSync(homeDir, { recursive: true });
-	writeAgentInstallState(agentDir);
-	writeRuntimeLayout(runtimeDir);
-	// Add node-compile-cache (all four allowed top-level entries present)
-	mkdirSync(join(runtimeDir, "node-compile-cache"), { recursive: true });
-	writeRuntimeMarker(runtimeDir);
-
-	const { result, output } = runUninstall(
-		["--dry-run", "--agent-dir", agentDir, "--bin-dir", binDir],
-		homeDir,
-	);
-	assert.equal(result.status, 0, `dry-run failed:\n${output}`);
-	// All four allowed entries present: must plan rm -rf (not SKIP)
-	assert.match(output, /rm -rf/, "runtime with bin+lib+node-compile-cache+marker must be planned for rm -rf");
-	assert.doesNotMatch(output, /leave it/i, "exclusivity allow-list must not trip on marker dotfile");
 });
